% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{AICc}
\alias{AICc}
\title{Akaike Corrected Information Criterion}
\usage{
AICc(object, ...)
}
\arguments{
\item{object}{a fitted model object for which there exists a logLik method
to extract the corresponding log-likelihood.}

\item{...}{optionally more fitted model objects.}
}
\value{
If just one object is provided, a numeric value with the corresponding AICC value.

If multiple objects are provided, a data.frame with rows corresponding to the objects and
columns representing the number of parameters in the model (df) and the AICC.
}
\description{
Akaike Corrected Information Criterion
}
