% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacopula.R
\name{arma2dvine}
\alias{arma2dvine}
\title{Transform an armacopula into a dvinecopula or dvinecopula2 object}
\usage{
arma2dvine(object)
}
\arguments{
\item{object}{an object of class \linkS4class{armacopula}.}
}
\value{
An object of class \linkS4class{dvinecopula} (for AR copulas)
or class \linkS4class{dvinecopula2} (for MA or ARMA copulas).
}
\description{
Transform an armacopula into a dvinecopula or dvinecopula2 object
}
\examples{
arma2dvine(armacopula(list(ar = 0.5, ma = 0.4)))
}
