% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacopula.R
\name{armacopula}
\alias{armacopula}
\title{Constructor function for ARMA copula process}
\usage{
armacopula(pars = list(ar = 0, ma = 0))
}
\arguments{
\item{pars}{list consisting of vector of AR parameters named `ar`
and vector of MA parameters named `ma`.}
}
\value{
An object of class \linkS4class{armacopula}.
}
\description{
Constructor function for ARMA copula process
}
\examples{
armacopula(list(ar = 0.5, ma = 0.4))
}
