% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{armafit2dvine}
\alias{armafit2dvine}
\title{Transform a fitted armacopula into a fitted dvinecopula or dvinecopula2 object}
\usage{
armafit2dvine(object)
}
\arguments{
\item{object}{an object of class \linkS4class{tscopulafit} in which the copula is
of class \linkS4class{armacopula}.}
}
\value{
An object of class \linkS4class{tscopulafit} in which the copula is
a \linkS4class{dvinecopula} (for fitted AR copulas)
or class \linkS4class{dvinecopula2} (for fitted MA or ARMA copulas).
}
\description{
Transform a fitted armacopula into a fitted dvinecopula or dvinecopula2 object
}
