% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacopula.R
\name{armavec2list}
\alias{armavec2list}
\title{Turn vector of ARMA parameters into list}
\usage{
armavec2list(theta, order)
}
\arguments{
\item{theta}{vector of ARMA model parameters}

\item{order}{order of model}
}
\value{
a list containing ARMA parameters in components ar and ma
}
\description{
Turn vector of ARMA parameters into list
}
