% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdvinecopula.R
\name{auto_dvine}
\alias{auto_dvine}
\title{Automatic estimation of stationary d-vine copula with Gaussian base family}
\usage{
auto_dvine(
  model,
  nreplace = 5,
  tautol = 1e-04,
  maxlag = Inf,
  ICtol = 0.5,
  nstrike = 3,
  criterion = "AIC",
  choices = c("gumbel", "clayton", "frank", "joe"),
  verbose = TRUE
)
}
\arguments{
\item{model}{an object of class \linkS4class{tscopulafit} such that \code{model@tscopula} is of class
\linkS4class{armacopula} or \linkS4class{sarmacopula}.}

\item{nreplace}{fixed number of copulas to replace}

\item{tautol}{tolerance for setting the effective maximum lag in list of copulas}

\item{maxlag}{fixed value for maximum lag}

\item{ICtol}{tolerance for deciding on a premature stop to copula replacement}

\item{nstrike}{number of negligible improvements before premature stop}

\item{criterion}{information criterion to be used (AIB, BIC or AICc)}

\item{choices}{vector of copula names to be used for replacements}

\item{verbose}{logical parameter specifying whether a verbose output is desired}
}
\value{
An object of class \linkS4class{tscopulafit} such that \code{model@tscopula} is of class
\linkS4class{sdvinecopula}.
}
\description{
This function begins with a model based exclusively on Gaussian pair copulas and
systematically replaces a finite number of copulas with non-Gaussian alternatives that
lower a specified information criterion.
}
\details{
The number of copulas to replace is specified by \code{nreplace} but the algorithm may stop
prematurely if the information criterion is barely lowered over \code{nstrike} consecutive steps.

This is a slow function to execute. Increasing \code{tautol} or setting a value for
\code{maxlag} will speed it up at the cost of some accuracy in the final model. The default
choices of copula are all one-parameter but the t copula can be added if desired.
For copulas without radial symmetry different rotations are tried to accommodate negative
and positive dependencies.
}
