% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_fit.R
\name{automatic_fit}
\alias{automatic_fit}
\title{Automatic IFM fit function for vt-D-vines by AIC}
\usage{
automatic_fit(
  y,
  margins = c("norm", "sst", "NIG", "weibull2", "burr2"),
  families = c(1, 13, 4, 5, 6),
  k = 4,
  fulcrum = FALSE,
  omethod = "Nelder-Mead",
  maxit = 5000
)
}
\arguments{
\item{y}{vector or time series of data to be used for fitting}

\item{margins}{a vector of marginal distributions to consider (selected using largest likelihood)}

\item{families}{a vector of families to consider for each pair copula of the d-vine}

\item{k}{maximum number of generalized lags to consider}

\item{fulcrum}{a pre-specified fulcrum value. If FALSE, automatic selection using profile likelihood is performed}

\item{omethod}{optimisation method for the marginal and final fit}

\item{maxit}{maximum number of iterations for omethod}
}
\value{
An object of class \linkS4class{tscopulafit}.
}
\description{
Automatic IFM fit function for vt-D-vines by AIC
}
