% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_fit.R
\name{automatic_step}
\alias{automatic_step}
\title{Automatic IFM fit function for vt-D-vines by stepwise data transformation}
\usage{
automatic_step(
  y,
  margins = c("norm", "sst", "NIG", "weibull2", "burr2"),
  families = c(1, 13, 4, 5, 6),
  k = 4,
  fulcrum = FALSE
)
}
\arguments{
\item{y}{vector or time series of data to be used for fitting}

\item{margins}{a vector of marginal distributions to consider (selected using largest likelihood)}

\item{families}{a vector of families to consider for each pair copula of the D-vine}

\item{k}{maximum number of generalized lags to consider}

\item{fulcrum}{a pre-specified fulcrum value. If FALSE, automatic selection using profile likelihood is performed}
}
\value{
An object of class \linkS4class{tscopula}.
}
\description{
Automatic IFM fit function for vt-D-vines by stepwise data transformation
}
