% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bitcoin}
\alias{bitcoin}
\title{Bitcoin price data 2016-19}
\format{
An object of class \code{"xts"}.
}
\usage{
data(bitcoin)
}
\description{
Time series of Bitcoin closing prices from 31 December 2015 to 31 December 2019 (1044 values).
This permits the calculation of 4 calendar years of returns.
}
\examples{
data(bitcoin)
plot(bitcoin)
X <- (diff(log(bitcoin))[-1]) * 100
plot(X)
}
\keyword{datasets}
