% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacopula.R
\name{coef,armacopula-method}
\alias{coef,armacopula-method}
\title{Coef Method for ARMA copula Class}
\usage{
\S4method{coef}{armacopula}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{armacopula}.}
}
\value{
parameters of ARMA copula model
}
\description{
Coef Method for ARMA copula Class
}
