% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula.R
\name{coef,dvinecopula-method}
\alias{coef,dvinecopula-method}
\title{Coef Method for dvinecopula Class}
\usage{
\S4method{coef}{dvinecopula}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{dvinecopula}.}
}
\value{
parameters of tscopula model
}
\description{
Coef Method for dvinecopula Class
}
