% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{coef,dvinecopula2-method}
\alias{coef,dvinecopula2-method}
\title{Coef Method for dvinecopula2 Class}
\usage{
\S4method{coef}{dvinecopula2}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{dvinecopula2}.}
}
\value{
parameters of tscopula model
}
\description{
Coef Method for dvinecopula2 Class
}
