% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{coef,margin-method}
\alias{coef,margin-method}
\title{Coef Method for margin Class}
\usage{
\S4method{coef}{margin}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{margin}.}
}
\value{
Vector of coefficients of marginal distribution.
}
\description{
Coef Method for margin Class
}
