% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_objects.R
\name{coef,swncopula-method}
\alias{coef,swncopula-method}
\title{Coef Method for Strict White Noise Copula}
\usage{
\S4method{coef}{swncopula}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{swncopula}.}
}
\value{
Null object
}
\description{
Coef Method for Strict White Noise Copula
}
