% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_models.R
\name{coef,tscm-method}
\alias{coef,tscm-method}
\title{Coefficient Method for tscm Class}
\usage{
\S4method{coef}{tscm}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{tscm}.}
}
\value{
vector of coefficients of model.
}
\description{
Coefficient Method for tscm Class
}
