% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{coef,tscopulafit-method}
\alias{coef,tscopulafit-method}
\title{Coefficients for tscopulafit Class}
\usage{
\S4method{coef}{tscopulafit}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{tscopulafit}.}
}
\value{
Coefficients of fitted model.
}
\description{
Coefficients for tscopulafit Class
}
