% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtransforms.R
\name{coef,Vtransform-method}
\alias{coef,Vtransform-method}
\title{Coef Method for Vtransform Class}
\usage{
\S4method{coef}{Vtransform}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{Vtransform}.}
}
\value{
Parameters of Vtransform model.
}
\description{
Coef Method for Vtransform Class
}
