% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_vtscopula.R
\name{coef,vtscopula-method}
\alias{coef,vtscopula-method}
\title{Coef Method for vtscopula Class}
\usage{
\S4method{coef}{vtscopula}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{vtscopula}.}
}
\value{
parameters of vtscopula model
}
\description{
Coef Method for vtscopula Class
}
