% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_models.R
\name{coerce,tscopulafit,tscmfit-method}
\alias{coerce,tscopulafit,tscmfit-method}
\title{Convert tscopulafit Object to be tscmfit Object}
\usage{
\S4method{coerce}{tscopulafit,tscmfit}(from, to = "tscmfit", strict = TRUE)
}
\arguments{
\item{from}{a \linkS4class{tscopulafit} object.}

\item{to}{a \linkS4class{tscmfit} object.}

\item{strict}{logical variable stating whether strict coercion should be enforced.}
}
\value{
A \linkS4class{tscmfit} object.
}
\description{
Convert tscopulafit Object to be tscmfit Object
}
