% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpi}
\alias{cpi}
\title{CPI inflation data 1959-2020}
\format{
An object of class \code{"xts"}.
}
\usage{
data(cpi)
}
\description{
Time series of US quarterly CPI (consumer price index) data Q4 1959 to Q4 2020 (245 values) for studying inflation.
These data were sourced from the OECD webpage and represent the total `perspective' on inflation, including food and energy.
They have been based to have a value of 100 in 2015.
}
\examples{
data(cpi)
plot(cpi)
X <- (diff(log(cpi))[-1]) * 100
plot(X)
}
\keyword{datasets}
