% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{dburr2}
\alias{dburr2}
\title{Density of burr2}
\usage{
dburr2(
  x,
  sh1 = 1,
  sh1b = 1,
  sc1 = 1,
  sh2 = 1,
  sh2b = 1,
  sc2 = 1,
  p0 = 0,
  log = F
)
}
\arguments{
\item{x}{variable}

\item{sh1}{shape parameter of right branch}

\item{sh1b}{second shape parameter of right branch}

\item{sc1}{scale parameter of right branch}

\item{sh2}{shape parameter of left branch}

\item{sh2b}{second shape parameter of left branch}

\item{sc2}{scale parameter of left branch}

\item{p0}{mixing parameter}
}
\value{

}
\description{
Density of burr2
}
\keyword{internal}
