% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_vtarma.R
\name{dcondvtarma}
\alias{dcondvtarma}
\title{Conditional density of VT-ARMA process}
\usage{
dcondvtarma(x, tscmmod, armamean, armasd)
}
\arguments{
\item{x}{vector of points at which density should be calculated.}

\item{tscmmod}{an object of class \linkS4class{tscmfit} based on underlying copula
of class \linkS4class{armacopula}.}

\item{armamean}{conditional mean of underlying ARMA process.}

\item{armasd}{conditional standard deviation of underlying ARMA process.}
}
\value{
vector with same length as x.
}
\description{
Conditional density of VT-ARMA process
}
\keyword{internal}
