% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{ddoubleweibull}
\alias{ddoubleweibull}
\title{Density of Double Weibull Distribution}
\usage{
ddoubleweibull(x, mu = 0, shape = 1, scale = 1, log = FALSE)
}
\arguments{
\item{x}{argument of density}

\item{mu}{location parameter}

\item{shape}{shape parameter}

\item{scale}{scale parameter}

\item{log}{flag for log density}
}
\value{

}
\description{
Density of Double Weibull Distribution
}
\keyword{internal}
