% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{dlaplace}
\alias{dlaplace}
\title{Density of Laplace Distribution}
\usage{
dlaplace(x, mu = 0, scale = 1, log = FALSE)
}
\arguments{
\item{x}{argument of density}

\item{mu}{location parameter}

\item{scale}{scale parameter}

\item{log}{flag for log density}
}
\value{

}
\description{
Density of Laplace Distribution
}
\keyword{internal}
