% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{dlgamma2}
\alias{dlgamma2}
\title{Density of lgamma2}
\usage{
dlgamma2(x, sh1 = 1, r1 = 2, sh2 = 1, r2 = 2, p0 = 0, log = F)
}
\arguments{
\item{x}{variable}

\item{sh1}{shapelog parameter of right branch}

\item{r1}{ratelog parameter of right branch}

\item{sh2}{shapelog parameter of left branch}

\item{r2}{ratelog parameter of left branch}

\item{p0}{mixing parameter}
}
\value{

}
\description{
Density of lgamma2
}
\keyword{internal}
