% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{dmarg}
\alias{dmarg}
\title{Compute density of marginal model}
\usage{
dmarg(x, y, log = FALSE)
}
\arguments{
\item{x}{an object of class \linkS4class{margin}.}

\item{y}{vector of values for which density should be computed.}

\item{log}{logical variable specifying whether log density should be
returned.}
}
\value{
A vector of values for the density.
}
\description{
Compute the density function of the marginal model.
}
\examples{
margmod <- margin("norm", pars = c(mean = 0, sd = 1))
dmarg(margmod, c(-2, 0, 2), log = TRUE)
}
