% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{dNIG}
\alias{dNIG}
\title{Density of NIG}
\usage{
dNIG(x, alphabar, gamma, mu, sigma, log = FALSE)
}
\arguments{
\item{x}{variable}

\item{log}{}
}
\value{

}
\description{
Density of NIG
}
\keyword{internal}
