% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{doubleweibull}
\alias{doubleweibull}
\alias{ddoubleweibull}
\alias{pdoubleweibull}
\alias{qdoubleweibull}
\alias{rdoubleweibull}
\title{Double Weibull distribution}
\usage{
ddoubleweibull(x, mu = 0.05, shape = 1, scale = 1, log = FALSE)

pdoubleweibull(q, mu = 0.05, shape = 1, scale = 1)

qdoubleweibull(p, mu = 0.05, shape = 1, scale = 1)

rdoubleweibull(n, mu = 0.05, shape = 1, scale = 1)
}
\arguments{
\item{x}{vector of values}

\item{mu}{location parameter}

\item{shape}{shape parameter}

\item{scale}{scale parameter}

\item{log}{flag for log density}

\item{q}{vector of quantiles}

\item{p}{vector of probabilities}

\item{n}{number of observations}
}
\description{
Double Weibull distribution
}
