% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{dsdoubleweibull}
\alias{dsdoubleweibull}
\title{Density of Skew Double Weibull Distribution}
\usage{
dsdoubleweibull(x, mu = 0, shape = 1, scale = 1, gamma = 1, log = FALSE)
}
\arguments{
\item{x}{argument of density}

\item{mu}{location parameter}

\item{shape}{shape parameter}

\item{scale}{scale parameter}

\item{gamma}{skewness parameter}

\item{log}{flag for log density}
}
\value{
A vector of values for the density.
}
\description{
Density of Skew Double Weibull Distribution
}
\keyword{internal}
