% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{dslaplace}
\alias{dslaplace}
\title{Density of Skew Laplace Distribution}
\usage{
dslaplace(x, mu = 0, scale = 1, gamma = 1, log = FALSE)
}
\arguments{
\item{x}{argument of density}

\item{mu}{location parameter}

\item{scale}{scale parameter}

\item{log}{flag for log density}

\item{skewness}{parameter}
}
\value{

}
\description{
Density of Skew Laplace Distribution
}
\keyword{internal}
