% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{dsst}
\alias{dsst}
\title{Density of Skewed Student t}
\usage{
dsst(x, df, gamma, mu, sigma, log = FALSE)
}
\arguments{
\item{x}{variable}

\item{df}{degrees of freedom}

\item{gamma}{skewness parameter}

\item{mu}{location parameter}

\item{sigma}{scale parameter}

\item{log}{}
}
\value{
A vector of values for the density.
}
\description{
Density of Skewed Student t
}
\keyword{internal}
