% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{dst}
\alias{dst}
\title{Density of Student t}
\usage{
dst(x, df, mu, sigma, log = FALSE)
}
\arguments{
\item{x}{variable}

\item{df}{degrees of freedom}

\item{mu}{location parameter}

\item{sigma}{scale parameter}

\item{log}{}
}
\value{
A vector of values for the density.
}
\description{
Density of Student t
}
\keyword{internal}
