% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula.R
\docType{class}
\name{dvinecopula-class}
\alias{dvinecopula-class}
\title{D-vine Copula Processes}
\description{
Class of objects for d-vine copula processes.
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the d-vine copula process.}

\item{\code{modelspec}}{list containing the family, number of parameters and rotations}

\item{\code{pars}}{list comprising of the parameters.}
}}

