% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\docType{class}
\name{dvinecopula2-class}
\alias{dvinecopula2-class}
\title{Type 2 D-vine Copula Processes}
\value{

}
\description{
Class of objects for d-vine copula processes.
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the d-vine copula process.}

\item{\code{modelspec}}{list containing the family, rotation, and name of KPACF}

\item{\code{pars}}{list comprising of the parameters.}
}}

