% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{dvinecopula2_objective}
\alias{dvinecopula2_objective}
\title{Objective function for dvinecopula2 process}
\usage{
dvinecopula2_objective(theta, modelspec, u)
}
\arguments{
\item{theta}{parameters of kpacf}

\item{modelspec}{list specifying model}

\item{u}{data}
}
\value{
Value of objective function at parameters.
}
\description{
Objective function for dvinecopula2 process
}
\keyword{internal}
