% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula3.R
\name{dvinecopula3_objective}
\alias{dvinecopula3_objective}
\title{Objective function for dvinecopula3 process}
\usage{
dvinecopula3_objective(theta, modelspec, u)
}
\arguments{
\item{theta}{parameters of kpacf}

\item{modelspec}{list specifying model}

\item{u}{data}
}
\value{
Value of objective function at parameters.
}
\description{
Objective function for dvinecopula3 process
}
\keyword{internal}
