% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopulavt.R
\name{dvinecopulavt}
\alias{dvinecopulavt}
\title{Constructor function for dvinecopulavt process}
\usage{
dvinecopulavt(
  family = "joe",
  rotation = 0,
  kpacf = "kpacf_arma",
  pars = list(ar = 0.1, ma = 0),
  vt1 = Vlinear(0.5),
  vt2 = Vlinear(0.5),
  tautol = 1e-04,
  maxlag = Inf,
  V2override = FALSE
)
}
\arguments{
\item{family}{family name}

\item{rotation}{a scalar specifying the rotation (default is 0)}

\item{kpacf}{a character string giving the name of the Kendall pacf}

\item{pars}{a list containing the parameters of the model}

\item{vt1}{first v-transform}

\item{vt2}{second v-transform}

\item{tautol}{scalar value at which kpacf is truncated}

\item{maxlag}{a scalar which can be used to force a given value for maximum lag}

\item{V2override}{logical variable stating whether 2-parameter v-transform
should be permitted}
}
\value{
An object of class \linkS4class{dvinecopulavt}.
}
\description{
This function sets up a stationary d-vine process of finite or infinite order based on a single
inverse-v-transformed copula family from a subset of those that can be implemented using
\code{\link[rvinecopulib]{bicop_dist}} in the \code{rvinecopulib} package.
}
\details{
The permitted choices of base copula family are currently Joe, Gumbel, Frank or Clayton survival. If
Clayton is chosen, the \code{rotation} argument must be set to 180, while if Joe or Gumbel are chosen, the
\code{rotation} argument must be zero (which is the default); any other options will return an error

The copulas are parameterized using the Kendall partial autocorrelation function (kpacf)
of the base copula sequence specified
by the \code{kpacf} argument. The default choice is the kpacf of a standard ARMA process which is
implemented in the function \code{\link{kpacf_arma}}. The parameters
of the kpacf should be set as a list using the \code{pars} argument; the required parameters should usually
be clear from the documentation of the chosen kpacf function and must be correctly named.

The arguments \code{vt1} and \code{vt2} are used to enter two parametric v-transforms which may be created, for example,
by \code{\link{Vlinear}} or \code{\link{V2p}}. However, the latter is very slow and the
variable \code{V2override} has to be set to \code{TRUE} if you want to include 2-parameter
v-transforms. While fitting is possible, residual analysis and simulation are almost always
prohibitively slow.

For data showing stochastic volatility, we expect positive serial dependencies in the base copula sequence.
For this reason, we do not consider models where the kpacf takes negative values.

In practice, the sequence of base copulas will be truncated at the last copula for which the kpacf exceeds \code{tautol}.
The \code{maxlag} parameter is typically used to force the truncation to take place at a lower lag (to increase speed).
This can also be achieved by increasing the value of \code{tautol}.
}
\examples{
dvinecopulavt(family = "joe", kpacf = "kpacf_arma",
pars = list(ar = 0.95, ma = -0.85), maxlag = 30)
}
