% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopulavt.R
\name{dvinecopulavt_objective}
\alias{dvinecopulavt_objective}
\title{Objective function for dvinecopulavt process}
\usage{
dvinecopulavt_objective(theta, modelspec, u)
}
\arguments{
\item{theta}{parameters of kpacf}

\item{modelspec}{list specifying model}

\item{u}{data}
}
\value{
Value of objective function at parameters.
}
\description{
Objective function for dvinecopulavt process
}
\keyword{internal}
