% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{edf}
\alias{edf}
\title{Construct empirical margin}
\usage{
edf()
}
\value{
An object of class \linkS4class{margin} signifying an empirical distribution function.
}
\description{
Construct empirical margin
}
