% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdvinecopula.R
\name{effective_kmax}
\alias{effective_kmax}
\title{Find effective maximum lag}
\usage{
effective_kmax(tau, modelspec)
}
\arguments{
\item{tau}{values of Kendall partial autocorrelation function.}

\item{modelspec}{sdvinecopula model specification}
}
\value{
a value for the effective maximum lag floored at one.
}
\description{
Find effective maximum lag
}
\keyword{internal}
