% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarmacopula.R
\name{expand_ar}
\alias{expand_ar}
\title{Expand AR coefficients to include SAR coefficients of SARMA model}
\usage{
expand_ar(ar, sar, period)
}
\arguments{
\item{ar}{vector of AR coefficients}

\item{sar}{vector of SAR coefficients}

\item{period}{period of SARMA model}
}
\value{
vector of AR coefficients in equivalent ARMA model
}
\description{
Expand AR coefficients to include SAR coefficients of SARMA model
}
\keyword{internal}
