% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarmacopula.R
\name{expand_ma}
\alias{expand_ma}
\title{Expand MA coefficients to include SMA coefficients of SARMA model}
\usage{
expand_ma(ma, sma, period)
}
\arguments{
\item{ma}{vector of MA coefficients}

\item{sma}{vector of SMA coefficients}

\item{period}{period of SARMA model}
}
\value{
vector of MA coefficients in equivalent ARMA model
}
\description{
Expand MA coefficients to include SMA coefficients of SARMA model
}
\keyword{internal}
