% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{fit}
\alias{fit}
\title{Generic for estimating time series models}
\usage{
fit(x, y, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{y}{a vector or time series of data.}

\item{...}{further arguments to be passed on.}
}
\value{
An object of the fitted model class.
}
\description{
Methods are available for objects of class \linkS4class{tscopulaU},
\linkS4class{vtscopula}, \linkS4class{tscopulafit}, \linkS4class{margin} and
\linkS4class{tscm}.
}
