% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_vtscopula.R
\name{fitFULLb}
\alias{fitFULLb}
\title{Fit tscm Jointly}
\usage{
fitFULLb(x, y, tsoptions, control)
}
\arguments{
\item{x}{an object of class \linkS4class{tscm}.}

\item{y}{a vector or time series of data.}

\item{tsoptions}{list of variables}

\item{control}{list of control parameters to be passed to the
\code{\link[stats]{optim}} function.}
}
\value{
an object of class \linkS4class{tscmfit}.
}
\description{
Fit tscm Jointly
}
\keyword{internal}
