% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{gauss}
\alias{gauss}
\alias{dgauss}
\alias{pgauss}
\alias{qgauss}
\alias{rgauss}
\title{Gaussian distribution}
\usage{
dgauss(x, mu = 0, sigma = 1, log = FALSE)

pgauss(q, mu = 0, sigma = 1)

qgauss(p, mu = 0, sigma = 1)

rgauss(n, mu = 0, sigma = 1)
}
\arguments{
\item{x}{vector of values.}

\item{mu}{location parameter.}

\item{sigma}{scale parameter.}

\item{log}{flag for log density.}

\item{q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
A vector of density, distribution function, quantile or random values.
}
\description{
Gaussian distribution
}
