% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{gauss0}
\alias{gauss0}
\alias{dgauss0}
\alias{pgauss0}
\alias{qgauss0}
\alias{rgauss0}
\title{Centred Gaussian distribution}
\usage{
dgauss0(x, sigma = 1, log = FALSE)

pgauss0(q, sigma = 1)

qgauss0(p, sigma = 1)

rgauss0(n, sigma = 1)
}
\arguments{
\item{x}{vector of values.}

\item{sigma}{scale parameter.}

\item{log}{flag for log density.}

\item{q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
A vector of density, distribution function, quantile or random values.
}
\description{
Centred Gaussian distribution
}
