% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula.R
\name{get_tau}
\alias{get_tau}
\title{Calculate Kendall's tau values for pair copulas in d-vine copula}
\usage{
get_tau(vinemodel)
}
\arguments{
\item{vinemodel}{a \linkS4class{dvinecopula} object}
}
\value{
vector consisting of Kendall's tau values for each pair copula
}
\description{
Calculate Kendall's tau values for pair copulas in d-vine copula
}
\examples{
mixmod <- dvinecopula(family = c("gumbel", "gauss", "joe", "clayton"), pars = list(1.5, -0.6, 1.6, 2.1))
get_tau(mixmod)
}
