% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tools.R
\name{getBiCop}
\alias{getBiCop}
\title{Extract BiCop objects from Fitted tscopula}
\usage{
getBiCop(fit)
}
\arguments{
\item{fit}{an object of class \linkS4class{tscopulafit} or \linkS4class{tscmfit}.}
}
\value{
a list of BiCop objects.
}
\description{
Extract BiCop objects from Fitted tscopula
}
