% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_vtscopula.R
\name{getparlist}
\alias{getparlist}
\title{Extract parameters of tscopula}
\usage{
getparlist(object)
}
\arguments{
\item{object}{an object of class \linkS4class{tscopula}.}
}
\value{
A list of parameters.
}
\description{
Extract parameters of tscopula
}
\keyword{internal}
