% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopulavt.R
\name{getvt}
\alias{getvt}
\title{Helper function for reparameterizing v-transforms}
\usage{
getvt(vt, theta, number)
}
\arguments{
\item{vt}{a v-transform}

\item{theta}{vector of parameters containing parameters of v-transform}

\item{number}{number of v-transform (1 or 2)}
}
\value{
a v-transform
}
\description{
Helper function for reparameterizing v-transforms
}
\keyword{internal}
