% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{gh}
\alias{gh}
\alias{pgh}
\alias{qgh}
\alias{dgh}
\alias{rgh}
\title{Generalized hyperbolic distribution}
\usage{
pgh(q, lambda = 0, shape = 1, gamma = 0, mu = 0, sigma = 1)

qgh(p, lambda, shape = 1, gamma = 0, mu = 0, sigma = 1)

dgh(x, lambda, shape = 1, gamma = 0, mu = 0, sigma = 1, log = FALSE)

rgh(n, lambda, shape = 1, gamma = 0, mu = 0, sigma = 0)
}
\arguments{
\item{q}{vector of quantiles.}

\item{lambda}{overarching shape parameter.}

\item{shape}{shape parameter.}

\item{gamma}{skewness parameter.}

\item{mu}{location parameter.}

\item{sigma}{scale parameter.}

\item{p}{vector of probabilities.}

\item{x}{vector of values.}

\item{log}{flag for log density.}

\item{n}{number of observations.}
}
\description{
Generalized hyperbolic distribution
}
