% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_margins.R
\name{Ghyp}
\alias{Ghyp}
\alias{pGhyp}
\alias{qGhyp}
\alias{dGhyp}
\alias{rGhyp}
\title{Generalized hyperbolic distribution}
\usage{
pGhyp(q, lambda = 0, shape = 1, gamma = 0, mu = 0, sigma = 1)

qGhyp(p, lambda, shape = 1, gamma = 0, mu = 0, sigma = 1)

dGhyp(x, lambda, shape = 1, gamma = 0, mu = 0, sigma = 1, log = FALSE)

rGhyp(n, lambda, shape = 1, gamma = 0, mu = 0, sigma = 0)
}
\arguments{
\item{q}{vector of quantiles}

\item{lambda}{overarching shape parameter}

\item{shape}{shape parameter}

\item{gamma}{skewness parameter}

\item{mu}{location parameter}

\item{sigma}{scale parameter}

\item{p}{vector of probabilities}

\item{x}{vector of values}

\item{n}{number of observations}
}
\description{
Generalized hyperbolic distribution
}
