% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{glag,tscopulafit-method}
\alias{glag,tscopulafit-method}
\title{glag Method for tscopulafit Class}
\usage{
\S4method{glag}{tscopulafit}(x, lagmax = 20, glagplot = FALSE)
}
\arguments{
\item{x}{an object of class \linkS4class{tscopulafit}.}

\item{glagplot}{logical value indicating generalized lag plot.}

\item{maxlag}{maximum value for lag.}
}
\value{
vector consisting of Kendall's tau values for each pair copula
}
\description{
glag Method for tscopulafit Class
}
