% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{glag}
\alias{glag}
\title{Generalized lagging function}
\usage{
glag(x, lagmax = 20, glagplot = FALSE)
}
\arguments{
\item{x}{an object of class \linkS4class{tscopulafit}.}

\item{lagmax}{maximum value for lag.}

\item{glagplot}{logical value indicating generalized lag plot.}
}
\value{
If \code{glagplot} is \code{TRUE} a list of generalized lagged datasets
of maximum length 9 is returned to facilitate a generalized lagplot.
If \code{glagplot} is \code{FALSE} a vector of length \code{lagmax} containing
the Kendall rank correlations for the generalized lagged datasets is returned.
}
\description{
Generalized lagging function
}
