% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacopula.R
\name{glag_for_armafit}
\alias{glag_for_armafit}
\title{Generalized lagging for fitted armacopula objects}
\usage{
glag_for_armafit(copula, data, lagmax, glagplot = FALSE)
}
\arguments{
\item{copula}{an armacopula object.}

\item{data}{the data to which copula is fitted.}

\item{lagmax}{the maximum lag value.}
}
\description{
Generalized lagging for fitted armacopula objects
}
\keyword{internal}
