% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula.R
\name{glag_for_dvinecopula}
\alias{glag_for_dvinecopula}
\title{Generalized lagging for fitted dvinecopula objects}
\usage{
glag_for_dvinecopula(copula, data, lagmax, glagplot = FALSE)
}
\arguments{
\item{copula}{a dvinecopula object}

\item{data}{the data to which copula is fitted}

\item{lagmax}{the maximum lag value.}

\item{glagplot}{logical value indicating generalized lag plot.}
}
\value{
If \code{glagplot} is \code{TRUE} a list of generalized lagged datasets
of maximum length 9 is returned to facilitate a generalized lagplot.
If \code{glagplot} is \code{FALSE} a vector of length \code{lagmax} containing
the Kendall rank correlations for the generalized lagged datasets is returned.
}
\description{
Generalized lagging for fitted dvinecopula objects
}
\keyword{internal}
