% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{glag_for_dvinecopula2}
\alias{glag_for_dvinecopula2}
\title{Generalized lagging for fitted dvinecopula2 objects}
\usage{
glag_for_dvinecopula2(copula, data, lagmax, glagplot = FALSE)
}
\arguments{
\item{copula}{a dvinecopula2 object}

\item{data}{the data to which copula is fitted}

\item{lagmax}{the maximum lag value.}

\item{glagplot}{logical value indicating generalized lag plot.}
}
\description{
Generalized lagging for fitted dvinecopula2 objects
}
\keyword{internal}
