% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula.R
\name{glag_for_dvinefit}
\alias{glag_for_dvinefit}
\title{Generalized lagging for fitted dvinecopula objects}
\usage{
glag_for_dvinefit(copula, data, glagplot = FALSE)
}
\arguments{
\item{copula}{a dvinecopula object}

\item{data}{the data to which copula is fitted}
}
\description{
Generalized lagging for fitted dvinecopula objects
}
\keyword{internal}
