% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{hyp}
\alias{hyp}
\alias{phyp}
\alias{qhyp}
\alias{dhyp}
\alias{rhyp}
\title{Hyperbolic distribution}
\usage{
phyp(q, shape = 1, gamma = 0, mu = 0, sigma = 1)

qhyp(p, shape = 1, gamma = 0, mu = 0, sigma = 1)

dhyp(x, shape = 1, gamma = 0, mu = 0, sigma = 1, log = FALSE)

rhyp(n, shape = 1, gamma = 0, mu = 0, sigma = 0)
}
\arguments{
\item{q}{vector of quantiles.}

\item{shape}{shape parameter.}

\item{gamma}{skewness parameter.}

\item{mu}{location parameter.}

\item{sigma}{scale parameter.}

\item{p}{vector of probabilities.}

\item{x}{vector of values.}

\item{log}{flag for log density.}

\item{n}{number of observations.}
}
\description{
Hyperbolic distribution
}
