% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_margins.R
\name{hypt}
\alias{hypt}
\alias{phypt}
\alias{qhypt}
\alias{dhypt}
\alias{rhypt}
\title{Hyperbolic t distribution}
\usage{
phypt(q, nu = 4, gamma = 0, mu = 0, sigma = 1)

qhypt(p, nu = 4, gamma = 0, mu = 0, sigma = 1)

dhypt(x, nu = 4, gamma = 0, mu = 0, sigma = 1, log = FALSE)

rhypt(n, nu = 4, gamma = 0, mu = 0, sigma = 0)
}
\arguments{
\item{q}{vector of quantiles}

\item{nu}{degree-of-freedom parameter}

\item{gamma}{skewness parameter}

\item{mu}{location parameter}

\item{sigma}{scale parameter}

\item{p}{vector of probabilities}

\item{x}{vector of values}

\item{n}{number of observations}
}
\description{
Hyperbolic t distribution
}
