% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_new_generics.R
\name{joint}
\alias{joint}
\title{New Generic for evaluating the joint density of a Time Series Models}
\usage{
joint(x, y, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{y}{a vector or time series of data.}

\item{...}{}
}
\value{
joint density
}
\description{
Methods are available for objects of class \linkS4class{tscopula},
\linkS4class{vtscopula}, \linkS4class{margin}
\linkS4class{tsc} and \linkS4class{gvtcopar}.
}
