% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{jsu}
\alias{jsu}
\alias{pjsu}
\alias{qjsu}
\alias{djsu}
\alias{rjsu}
\title{Johnson's SU distribution}
\usage{
pjsu(q, delta = 2, mu = 0, sigma = 1, gamma = 0)

qjsu(p, delta, mu, sigma, gamma)

djsu(x, delta, mu, sigma, gamma, log = FALSE)

rjsu(n, delta, mu, sigma, gamma)
}
\arguments{
\item{q}{vector of quantiles.}

\item{delta}{shape parameter.}

\item{mu}{location parameter.}

\item{sigma}{scale parameter.}

\item{gamma}{skewness parameter.}

\item{p}{vector of probabilities.}

\item{x}{vector of values.}

\item{log}{flag for log density.}

\item{n}{number of observations.}
}
\value{
A vector of density, distribution function, quantile or random values.
}
\description{
Johnson's SU distribution
}
