% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacopula.R
\name{kendall,armacopula-method}
\alias{kendall,armacopula-method}
\title{Calculate Kendall's tau values for armacopula model}
\usage{
\S4method{kendall}{armacopula}(x, lagmax = 20)
}
\arguments{
\item{x}{a \linkS4class{armacopula} object}

\item{maxlag}{maximum value of lag}
}
\value{
vector consisting of Kendall's tau values for each pair copula
}
\description{
Calculate Kendall's tau values for armacopula model
}
\examples{
mod <- armacopula(list(ar = 0.95, ma = -0.85))
kendall(mod)
}
