% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula.R
\name{kendall,dvinecopula-method}
\alias{kendall,dvinecopula-method}
\title{Calculate Kendall's tau values for pair copulas in d-vine copula}
\usage{
\S4method{kendall}{dvinecopula}(x, lagmax = NA)
}
\arguments{
\item{x}{a \linkS4class{dvinecopula} object}
}
\value{
vector consisting of Kendall's tau values for each pair copula
}
\description{
Calculate Kendall's tau values for pair copulas in d-vine copula
}
\examples{
mixmod <- dvinecopula(family = c("gumbel", "gauss"), pars = list(1.5, -0.6))
kendall(mixmod)
}
