% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_vtscopula.R
\name{kendall,vtscopula-method}
\alias{kendall,vtscopula-method}
\title{Calculate Kendall's tau values for vtscopula model}
\usage{
\S4method{kendall}{vtscopula}(x, lagmax = 20)
}
\arguments{
\item{x}{a \linkS4class{vtscopula} object}

\item{lagmax}{maximum value of lag}
}
\value{
vector consisting of Kendall's tau values for each pair copula
}
\description{
Calculate Kendall's tau values for vtscopula model
}
\examples{
mod <- vtscopula(armacopula(list(ar = 0.95, ma = -0.85)))
kendall(mod)
}
