% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_objects.R
\name{kendall}
\alias{kendall}
\title{New Generic for Kendall correlations}
\usage{
kendall(x, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{...}{further arguments to be passed on.}
}
\value{
A realization from the time series model.
}
\description{
Methods are available for objects of class \linkS4class{armacopula}
 and \linkS4class{dvinecopula}.
}
