% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{kpacf_arfima}
\alias{kpacf_arfima}
\title{KPACF of ARFIMA process}
\usage{
kpacf_arfima(k, theta)
}
\arguments{
\item{k}{number of lags.}

\item{theta}{list with components ar, ma and d specifying the ARFIMA parameters}
}
\value{
A vector of Kendall partial autocorrelations of length \code{k}.
}
\description{
KPACF of ARFIMA process
}
