% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{kpacf_arfima1}
\alias{kpacf_arfima1}
\title{KPACF of ARFIMA Process}
\usage{
kpacf_arfima1(k, theta)
}
\arguments{
\item{k}{vector of lags}

\item{theta}{list with components ar, ma and H specifying the ARFIMA parameters}
}
\value{
vector of Kendall partial autocorrelations for each lag k
}
\description{
KPACF of ARFIMA Process
}
