% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{kpacf_arma}
\alias{kpacf_arma}
\title{KPACF of ARMA process}
\usage{
kpacf_arma(k, theta)
}
\arguments{
\item{k}{number of lags.}

\item{theta}{list with components ar and ma specifying the ARMA parameters.}
}
\value{
A vector of Kendall partial autocorrelations of length \code{k}.
}
\description{
KPACF of ARMA process
}
