% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{kpacf_exp}
\alias{kpacf_exp}
\title{KPACF of Exponential Type}
\usage{
kpacf_exp(k, theta)
}
\arguments{
\item{k}{vector of lags}

\item{theta}{parameters of exponential decay function}
}
\value{
vector of Kendall partial autocorrelations for each lag k
}
\description{
KPACF of Exponential Type
}
