% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{kpacf_fbn}
\alias{kpacf_fbn}
\title{KPACF of fractional Brownian noise}
\usage{
kpacf_fbn(k, theta)
}
\arguments{
\item{k}{vector of lags}

\item{theta}{parameter of process}
}
\value{
vector of Kendall partial autocorrelations for each lag k
}
\description{
KPACF of fractional Brownian noise
}
