% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{kpacf_pow}
\alias{kpacf_pow}
\title{KPACF of Power Type}
\usage{
kpacf_pow(k, theta)
}
\arguments{
\item{k}{vector of lags}

\item{theta}{parameters of power decay function}
}
\value{
vector of Kendall partial autocorrelations for each lag k
}
\description{
KPACF of Power Type
}
