% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{kpacf_pow2}
\alias{kpacf_pow2}
\title{KPACF of Transformed Power Type}
\usage{
kpacf_pow2(k, theta)
}
\arguments{
\item{k}{vector of lags}

\item{theta}{parameters of power decay function}
}
\value{
vector of Kendall partial autocorrelations for each lag k
}
\description{
KPACF of Transformed Power Type
}
