% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{kpacf_sarma4}
\alias{kpacf_sarma4}
\title{KPACF of quarterly seasonal ARMA process}
\usage{
kpacf_sarma4(k, theta)
}
\arguments{
\item{k}{number of lags.}

\item{theta}{list with components ar, ma, sar and sma specifying the ARMA and seasonal ARMA parameters.}
}
\value{
A vector of Kendall partial autocorrelations of length \code{k}.
}
\description{
KPACF of quarterly seasonal ARMA process
}
