% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{laplace0}
\alias{laplace0}
\alias{dlaplace0}
\alias{plaplace0}
\alias{qlaplace0}
\alias{rlaplace0}
\title{Centred Laplace distribution}
\usage{
dlaplace0(x, scale = 1, log = FALSE)

plaplace0(q, scale = 1)

qlaplace0(p, scale = 1)

rlaplace0(n, scale = 1)
}
\arguments{
\item{x}{vector of values.}

\item{scale}{scale parameter.}

\item{log}{flag for log density.}

\item{q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
A vector of density, distribution function, quantile or random values.
}
\description{
Centred Laplace distribution
}
