% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{logLik,marginfit-method}
\alias{logLik,marginfit-method}
\title{logLik Method for marginfit Class}
\usage{
\S4method{logLik}{marginfit}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{marginfit}.}
}
\value{
an object of class logLik
}
\description{
logLik Method for marginfit Class
}
