% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_models.R
\name{logLik,tscmfit-method}
\alias{logLik,tscmfit-method}
\title{logLik Method for tscmfit Class}
\usage{
\S4method{logLik}{tscmfit}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{tscmfit}.}
}
\value{
An object of class logLik.
}
\description{
logLik Method for tscmfit Class
}
