% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{logLik,tscopulafit-method}
\alias{logLik,tscopulafit-method}
\title{logLik Method for tscopulafit Class}
\usage{
\S4method{logLik}{tscopulafit}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{tscopulafit}.}
}
\value{
an object of class logLik
}
\description{
logLik Method for tscopulafit Class
}
