% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tools.R
\name{logLik,uGARCHfit-method}
\alias{logLik,uGARCHfit-method}
\title{logLik Method for uGARCHfit Class}
\usage{
\S4method{logLik}{uGARCHfit}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{uGARCHfit}.}
}
\value{
an object of class logLik
}
\description{
logLik Method for uGARCHfit Class
}
