% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tools.R
\name{LRT,tscmfit,tscmfit-method}
\alias{LRT,tscmfit,tscmfit-method}
\title{LRT Method for tscmfit Class}
\usage{
\S4method{LRT}{tscmfit,tscmfit}(x, y)
}
\arguments{
\item{x, y}{objects of class \linkS4class{tscfit}.}
}
\value{
LRT between the models.
}
\description{
LRT Method for tscmfit Class
}
