% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tools.R
\name{LRT,tscopulafit,tscopulafit-method}
\alias{LRT,tscopulafit,tscopulafit-method}
\title{LRT Method for tscopulafit Class}
\usage{
\S4method{LRT}{tscopulafit,tscopulafit}(x, y)
}
\arguments{
\item{x, y}{objects of class \linkS4class{tscopulafit}.}
}
\value{
LRT between the models.
}
\description{
LRT Method for tscopulafit Class
}
