% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_new_generics.R
\name{LRT}
\alias{LRT}
\title{New Generic for carrying out a likelihood ratio test between two Time Series Models}
\usage{
LRT(x, y, ...)
}
\arguments{
\item{x, y}{objects of the model class.}

\item{...}{}
}
\value{
a likelihood ratio test result.
}
\description{
Methods are available for objects of class \linkS4class{tscopulafit},
\linkS4class{tscmfit}
}
