% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{margin}
\alias{margin}
\title{Constructor Function for margin}
\usage{
margin(name, pars = NULL, fixed = FALSE)
}
\arguments{
\item{name}{character string giving name of distribution}

\item{pars}{parameters of the distribution}

\item{fixed}{fixed parameters, for estimation purposes}
}
\value{
an object of class \linkS4class{margin}
}
\description{
Constructor Function for margin
}
\examples{
margin("sst")
margin("weibull2")
}
