% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\docType{class}
\name{marginfit-class}
\alias{marginfit-class}
\alias{logLik,marginfit-method}
\title{Fitted marginal model for time series}
\usage{
\S4method{logLik}{marginfit}(object)
}
\arguments{
\item{object}{an object of the class.}
}
\description{
Fitted marginal model for time series
}
\section{Methods (by generic)}{
\itemize{
\item \code{logLik}: logLik method for marginfit class
}}

\section{Slots}{

\describe{
\item{\code{margin}}{an object of class \linkS4class{margin}.}

\item{\code{data}}{numeric vector or time series of data.}

\item{\code{fit}}{a list containing details of the maximum likelihood fit.}
}}

