% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula.R
\name{mklist_dvine}
\alias{mklist_dvine}
\title{Make list of pair copulas for dvinecopula object}
\usage{
mklist_dvine(x)
}
\arguments{
\item{x}{an object of class dvinecopula}
}
\value{
a list of pair copulas
}
\description{
Make list of pair copulas for dvinecopula object
}
\keyword{internal}
