% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{mklist_dvine2}
\alias{mklist_dvine2}
\title{Make list of pair copulas for dvinecopula2 object}
\usage{
mklist_dvine2(x, maxlag)
}
\arguments{
\item{x}{an object of class dvinecopula2}

\item{maxlag}{maximum possible lag to consider}
}
\value{
a list of pair copulas
}
\description{
Make list of pair copulas for dvinecopula2 object
}
\keyword{internal}
