% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopulavt.R
\name{mklist_dvinevt}
\alias{mklist_dvinevt}
\title{Make list of pair copulas for dvinecopulavt object}
\usage{
mklist_dvinevt(x, maxlag)
}
\arguments{
\item{x}{an object of class dvinecopulavt}

\item{maxlag}{maximum possible lag to consider}
}
\value{
a list of pair copulas
}
\description{
Make list of pair copulas for dvinecopulavt object
}
\keyword{internal}
