% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdvinecopula.R
\name{mklist_sdvine}
\alias{mklist_sdvine}
\title{Make list of pair copulas for sdvinecopula object}
\usage{
mklist_sdvine(x, maxlag)
}
\arguments{
\item{x}{an object of class sdvinecopula}

\item{maxlag}{maximum possible lag to consider}
}
\value{
a list of pair copulas
}
\description{
Make list of pair copulas for sdvinecopula object
}
\keyword{internal}
