% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{nig}
\alias{nig}
\alias{pnig}
\alias{qnig}
\alias{dnig}
\alias{rnig}
\title{Normal inverse Gaussian distribution}
\usage{
pnig(q, shape = 1, gamma = 0, mu = 0, sigma = 1)

qnig(p, shape = 1, gamma = 0, mu = 0, sigma = 1)

dnig(x, shape = 1, gamma = 0, mu = 0, sigma = 1, log = FALSE)

rnig(n, shape = 1, gamma = 0, mu = 0, sigma = 0)
}
\arguments{
\item{q}{vector of quantiles.}

\item{shape}{shape parameter.}

\item{gamma}{skewness parameter.}

\item{mu}{location parameter.}

\item{sigma}{scale parameter.}

\item{p}{vector of probabilities.}

\item{x}{vector of values.}

\item{log}{flag for log density.}

\item{n}{number of observations.}
}
\description{
Normal inverse Gaussian distribution
}
