% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacopula.R
\name{non_invert}
\alias{non_invert}
\title{Check for Invertibility}
\usage{
non_invert(ma)
}
\arguments{
\item{ma}{}
}
\value{
A logical variable stating whether ARMA process is invertible.
}
\description{
Check for Invertibility
}
\keyword{internal}
