% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacopula.R
\name{non_stat}
\alias{non_stat}
\title{Check for causality of ARMA process}
\usage{
non_stat(ar)
}
\arguments{
\item{ar}{vector of autoregressive parameters}
}
\value{
A logical variable stating whether ARMA process is causal.
}
\description{
Check for causality of ARMA process
}
