% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{pacf2acf}
\alias{pacf2acf}
\title{Compute Autocorrelations from Partial Autocorrelations}
\usage{
pacf2acf(alpha)
}
\arguments{
\item{alpha}{vector of partial autocorrelation values}
}
\value{
vector of autocorrelation values
}
\description{
Compute Autocorrelations from Partial Autocorrelations
}
\examples{
alpha <- ARMAacf(ar = -0.9, ma = 0.8, lag.max = 50, pacf = TRUE)
rho <- pacf2acf(alpha)
}
