% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{pacf2ar}
\alias{pacf2ar}
\title{Compute autoregressive coefficients from partial autocorrelations}
\usage{
pacf2ar(alpha)
}
\arguments{
\item{alpha}{vector of partial autocorrelation values.}
}
\value{
A vector of autoregressive coefficients with same length as \code{alpha}.
}
\description{
Compute autoregressive coefficients from partial autocorrelations
}
\examples{
alpha <- ARMAacf(ar = -0.9, ma = 0.8, lag.max = 50, pacf = TRUE)
phi <- pacf2ar(alpha)
}
