% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{pburrgamma}
\alias{pburrgamma}
\title{Distribution Function of burrgamma}
\usage{
pburrgamma(q, sh1 = 1, sh1b = 1, sc1 = 1, sh2 = 1, sh2b = 1, sc2 = 1, p0 = 0)
}
\arguments{
\item{q}{quantile}

\item{sh1}{shape parameter of right branch}

\item{sh1b}{second shape parameter of right branch}

\item{sc1}{scale parameter of right branch}

\item{sh2}{shape parameter of left branch}

\item{sh2b}{second shape parameter of left branch}

\item{sc2}{scale parameter of left branch}

\item{p0}{mixing parameter}
}
\value{

}
\description{
Distribution Function of burrgamma
}
\keyword{internal}
