% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtransforms.R
\name{pcoincide}
\alias{pcoincide}
\title{Compute coincidence probability for v-transform}
\usage{
pcoincide(x)
}
\arguments{
\item{x}{an object of class \linkS4class{Vtransform}.}
}
\value{
The probability of coincidence.
}
\description{
Computes the probability that if we v-transform a uniform
random variable and then stochastically invert the
v-transform, we get back to the original value.
}
\examples{
pcoincide(Vlinear(delta = 0.4))
pcoincide(V3p(delta = 0.45, kappa = 0.5, xi = 1.3))
}
