% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{pdoubleweibull}
\alias{pdoubleweibull}
\title{Distribution Function of Double Weibull Distribution}
\usage{
pdoubleweibull(q, mu = 0, shape = 1, scale = 1)
}
\arguments{
\item{q}{argument of distribution function}

\item{mu}{location parameter}

\item{shape}{shape paramater}

\item{scale}{scale parameter}
}
\value{
A vector of values for the CDF.
}
\description{
Distribution Function of Double Weibull Distribution
}
\keyword{internal}
