% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{plaplace}
\alias{plaplace}
\title{Distribution Function of Laplace Distribution}
\usage{
plaplace(q, mu = 0, scale = 1)
}
\arguments{
\item{q}{argument of distribution function}

\item{mu}{location parameter}

\item{scale}{scale parameter}
}
\value{
A vector of values for the distribution function.
}
\description{
Distribution Function of Laplace Distribution
}
\keyword{internal}
