% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{plgamma2}
\alias{plgamma2}
\title{Distribution Function of lgamma2}
\usage{
plgamma2(q, sh1 = 1, r1 = 2, sh2 = 1, r2 = 2, p0 = 0)
}
\arguments{
\item{q}{quantile}

\item{sh1}{shapelog parameter of right branch}

\item{r1}{ratelog parameter of right branch}

\item{sh2}{shapelog parameter of left branch}

\item{r2}{ratelog parameter of left branch}

\item{p0}{mixing parameter}
}
\value{

}
\description{
Distribution Function of lgamma2
}
\keyword{internal}
