% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{plot,marginfit,missing-method}
\alias{plot,marginfit,missing-method}
\title{Plot method for marginfit class}
\usage{
\S4method{plot}{marginfit,missing}(x, bw = FALSE)
}
\arguments{
\item{x}{an object of class \linkS4class{marginfit}.}

\item{bw}{logical variable specifying whether black-white options should be chosen.}
}
\description{
Plot method for marginfit class
}
