% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{plot,tscopulafit,missing-method}
\alias{plot,tscopulafit,missing-method}
\title{Plot method for tscopulafit class}
\usage{
\S4method{plot}{tscopulafit,missing}(x, plottype = "residual", bw = FALSE, lagmax = 30)
}
\arguments{
\item{x}{an object of class \linkS4class{tscopulafit}.}

\item{plottype}{type of plot required.}

\item{bw}{logical variable specifying whether black-white options should be chosen.}

\item{lagmax}{maximum lag value for Kendall plots}
}
\description{
Plot method for tscopulafit class
}
\examples{
data <- sim(armacopula(list(ar = 0.5, ma = 0.4)), n = 1000)
fit <- fit(armacopula(list(ar = 0.5, ma = 0.4)), data)
plot(fit)
}
