% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacopula.R
\name{plot_armacopula}
\alias{plot_armacopula}
\title{Plot Function for armacopula Objects}
\usage{
plot_armacopula(copula, data, plotoption, bw)
}
\arguments{
\item{copula}{a fitted armacopula object}

\item{data}{the data to which copula is fitted}

\item{plotoption}{number giving plot choice}

\item{bw}{logical for black-white plot}
}
\value{

}
\description{
Plot Function for armacopula Objects
}
