% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula.R
\name{plot_dvinecopula}
\alias{plot_dvinecopula}
\title{Plot Function for dvinecopula Objects}
\usage{
plot_dvinecopula(copula, data, plotoption, bw)
}
\arguments{
\item{copula}{a fitted dvinecopula object}

\item{data}{the data to which copula is fitted}

\item{plotoption}{number giving plot choice}

\item{bw}{logical for black-white plot}
}
\description{
Plot Function for dvinecopula Objects
}
