% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{plot_dvinecopula2}
\alias{plot_dvinecopula2}
\title{Plot Function for dvinecopula2 Objects}
\usage{
plot_dvinecopula2(copula, data, plotoption, bw, klimit)
}
\arguments{
\item{copula}{a fitted dvinecopula2 object}

\item{data}{the data to which copula is fitted}

\item{plotoption}{number giving plot choice}

\item{bw}{logical for black-white plot}

\item{klimit}{maximum lag value for plots}
}
\value{

}
\description{
Plot Function for dvinecopula2 Objects
}
