% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_models.R
\name{plot_volprofile}
\alias{plot_volprofile}
\title{Plot function for volatility profile plot}
\usage{
plot_volprofile(x, bw)
}
\arguments{
\item{x}{an object of class \linkS4class{tscmfit}.}

\item{bw}{logical variable specifying whether black-white options should be chosen.}
}
\value{
No return value, generates plot.
}
\description{
Plot function for volatility profile plot
}
\keyword{internal}
