% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_models.R
\name{plot_volproxy}
\alias{plot_volproxy}
\title{Plot Function for Volatility Proxy Plot}
\usage{
plot_volproxy(x, bw)
}
\arguments{
\item{x}{an object of class \linkS4class{tscmfit}.}

\item{bw}{logical variable specifying whether black-white options should be chosen.}
}
\description{
Plot Function for Volatility Proxy Plot
}
