% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{pmarg}
\alias{pmarg}
\title{Compute CDF of Marginal Model}
\usage{
pmarg(x, q)
}
\arguments{
\item{x}{an object of class \linkS4class{margin}.}

\item{q}{vector of values at which CDF should be computed.}
}
\value{
A vector of values for the CDF.
}
\description{
Compute the cumulative distribution function of the marginal model.
}
\examples{
margmod <- margin("norm", pars = c(mean = 0, sd = 1))
pmarg(margmod, c(-2, 0, 2))
}
