% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{pNIG}
\alias{pNIG}
\title{Distribution Function of NIG}
\usage{
pNIG(q, alphabar = 1, gamma = 0, mu = 0, sigma = 1)
}
\arguments{
\item{q}{quantile}

\item{sigma}{}
}
\value{

}
\description{
Distribution Function of NIG
}
\keyword{internal}
