% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_models.R
\name{predict_empirical}
\alias{predict_empirical}
\title{Prediction function for tscm class with empirical margin}
\usage{
predict_empirical(object, data, x, type, qtype)
}
\arguments{
\item{object}{an object of the class.}

\item{data}{vector of past data values.}

\item{x}{vector of arguments of prediction function.}

\item{type}{type of prediction function ("df" for density, "qf" for quantile function
or "dens" for density).}

\item{qtype}{type of empirical quantile estimate.}
}
\description{
Prediction function for tscm class with empirical margin
}
\keyword{internal}
