% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{psdoubleweibull}
\alias{psdoubleweibull}
\title{Distribution Function of Skew Double Weibull Distribution}
\usage{
psdoubleweibull(q, mu = 0, shape = 1, scale = 1, gamma = 1)
}
\arguments{
\item{q}{argument of distribution function}

\item{mu}{location parameter}

\item{shape}{shape paramater}

\item{scale}{scale parameter}

\item{gamma}{skewness parameter}
}
\value{

}
\description{
Distribution Function of Skew Double Weibull Distribution
}
\keyword{internal}
