% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{psst}
\alias{psst}
\title{Distribution Function of Skewed Student t}
\usage{
psst(q, df = 10, gamma = 1, mu = 0, sigma = 1)
}
\arguments{
\item{q}{quantile}

\item{df}{degrees of freedom}

\item{gamma}{skewness parameter}

\item{mu}{location parameter}

\item{sigma}{scale parameter}
}
\value{
A vector of values for the CDF.
}
\description{
Distribution Function of Skewed Student t
}
\keyword{internal}
