% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{pst}
\alias{pst}
\title{Distribution Function of Student t}
\usage{
pst(q, df = 10, mu = 0, sigma = 1)
}
\arguments{
\item{q}{quantile}

\item{df}{degrees of freedom}

\item{mu}{location parameter}

\item{sigma}{scale parameter}
}
\value{

}
\description{
Distribution Function of Student t
}
\keyword{internal}
