% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{qdoubleweibull}
\alias{qdoubleweibull}
\title{Quantile Function of Double Weibull Distribution}
\usage{
qdoubleweibull(p, mu = 0, shape = 1, scale = 1)
}
\arguments{
\item{p}{probability}

\item{mu}{location parameter}

\item{shape}{shape parameter}

\item{scale}{scale parameter}
}
\value{

}
\description{
Quantile Function of Double Weibull Distribution
}
\keyword{internal}
