% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{qedf}
\alias{qedf}
\title{Quantiles of Empirical Distribution Function}
\usage{
qedf(p, data)
}
\arguments{
\item{p}{vector of probabilities at which to evaluate the quantiles
of the empirical distribution function.}

\item{data}{vector of data values specifying the empirical distribution
function}
}
\value{

}
\description{
Quantiles of Empirical Distribution Function
}
\examples{
data <- rnorm(100)
qedf(c(0.25, 0.5, 0.75), data)
}
