% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{qlaplace}
\alias{qlaplace}
\title{Quantile Function of Laplace Distribution}
\usage{
qlaplace(p, mu = 0, scale = 1)
}
\arguments{
\item{p}{probability}

\item{mu}{location parameter}

\item{scale}{scale parameter}
}
\value{
A vector of values for the quantile function.
}
\description{
Quantile Function of Laplace Distribution
}
\keyword{internal}
