% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{qNIG}
\alias{qNIG}
\title{Quantile Function of NIG}
\usage{
qNIG(p, alphabar, gamma, mu, sigma)
}
\arguments{
\item{p}{probability}

\item{sigma}{}
}
\value{

}
\description{
Quantile Function of NIG
}
\keyword{internal}
