% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{qsdoubleweibull}
\alias{qsdoubleweibull}
\title{Quantile Function of Skew Double Weibull Distribution}
\usage{
qsdoubleweibull(p, mu = 0, shape = 1, scale = 1, gamma = 1)
}
\arguments{
\item{p}{probability}

\item{mu}{location parameter}

\item{shape}{shape parameter}

\item{scale}{scale parameter}

\item{gamma}{skewness parameter}
}
\value{
A vector of values for the quantile function.
}
\description{
Quantile Function of Skew Double Weibull Distribution
}
\keyword{internal}
