% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{qslaplace}
\alias{qslaplace}
\title{Quantile Function of Skew Laplace Distribution}
\usage{
qslaplace(p, mu = 0, scale = 1, gamma = 1)
}
\arguments{
\item{p}{probability}

\item{mu}{location parameter}

\item{scale}{scale parameter}

\item{gamma}{skewness parameter}
}
\value{

}
\description{
Quantile Function of Skew Laplace Distribution
}
\keyword{internal}
