% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{qsst}
\alias{qsst}
\title{Quantile Function of Skewed Student t}
\usage{
qsst(p, df, gamma, mu, sigma)
}
\arguments{
\item{p}{probability}

\item{df}{degrees of freedom}

\item{gamma}{skewness parameter}

\item{mu}{location parameter}

\item{sigma}{scale parameter}
}
\value{
A vector of values for the quantile function.
}
\description{
Quantile Function of Skewed Student t
}
\keyword{internal}
