% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{qst}
\alias{qst}
\title{Quantile Function of Student t}
\usage{
qst(p, df, mu, sigma)
}
\arguments{
\item{p}{probability}

\item{df}{degrees of freedom}

\item{mu}{location parameter}

\item{sigma}{scale parameter}
}
\value{
A vector of values for the quantile function.
}
\description{
Quantile Function of Student t
}
\keyword{internal}
