% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{qweibull2}
\alias{qweibull2}
\title{Quantile Function of weibull2}
\usage{
qweibull2(p, sh1 = 1, sc1 = 1, sh2 = 1, sc2 = 1, p0 = 0)
}
\arguments{
\item{p}{probability}

\item{sh1}{shape parameter of right branch}

\item{sc1}{scale parameter of right branch}

\item{sh2}{shape parameter of left branch}

\item{sc2}{scale parameter of left branch}

\item{p0}{mixing parameter}
}
\value{

}
\description{
Quantile Function of weibull2
}
\keyword{internal}
