% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tools.R
\name{racf}
\alias{racf}
\title{Rank Autocorrelation Function}
\usage{
racf(
  data,
  method = "kendall",
  nlags = floor(5 * log10(length(data))),
  plot = TRUE,
  main = paste("Series", deparse(substitute(data)), sep = " "),
  xlab = "Lag",
  ylab = "ACF",
  lwd = 2
)
}
\arguments{
\item{data}{a vector of data to which the autocorrelation function is to be computed}

\item{method}{one of three ways to compute the acf: "kendall", "spearman" or "pearson"}

\item{nlags}{the number of lags considered}

\item{plot}{logical specifing whether to plot the acf or not}

\item{main}{title of the plot}

\item{xlab}{x-axis label of the plot}

\item{ylab}{y-axis label of the plot}

\item{lwd}{width of lines}
}
\value{
a vector containing the autocorrelation values
}
\description{
Rank Autocorrelation Function
}
\examples{
racf(runif(1000))
y <- sim(dvinecopula("gauss", list(0.5)))
racf(y)
}
