% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopulavt.R
\name{Rbackward}
\alias{Rbackward}
\title{Rosenblatt backward function with v-transforms}
\usage{
Rbackward(x, u, pcs, vt1, vt2)
}
\arguments{
\item{x}{vector argument of Rosenblatt functiom}

\item{u}{matrix of conditioning values. Number of rows
must be either 1 or same length as x. Number of columns
should not be much more than 15 (due to repeated recursive
calling)}

\item{pcs}{list of pair copulas}

\item{vt1}{first v-transform}

\item{vt2}{second v-transform}
}
\value{
vector of same length as x
}
\description{
Rosenblatt backward function with v-transforms
}
