% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula.R
\name{Rblatt}
\alias{Rblatt}
\title{Calculate Rosenblatt function}
\usage{
Rblatt(pc_list, data, x)
}
\arguments{
\item{pc_list}{a list of pair copulas.}

\item{data}{vector of past data values.}

\item{x}{vector of arguments of Rosenblatt function.}
}
\value{
a vector of values with same length as \code{x}.
}
\description{
Calculate Rosenblatt function
}
\keyword{internal}
