% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula.R
\name{Rblattdens}
\alias{Rblattdens}
\title{Calculate Rosenblatt density function}
\usage{
Rblattdens(pc_list, data, x)
}
\arguments{
\item{pc_list}{a list of pair copulas.}

\item{data}{vector of past data values.}

\item{x}{vector of arguments of Rosenblatt density function.}
}
\value{
a vector of values with same length as \code{x}.
}
\description{
Calculate Rosenblatt density function
}
\keyword{internal}
