% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{rdoubleweibull}
\alias{rdoubleweibull}
\title{Random Number Generation for Double Weibull Distribution}
\usage{
rdoubleweibull(n, mu = 0, shape = 1, scale = 1)
}
\arguments{
\item{n}{size}

\item{mu}{location parameter}

\item{shape}{shape parameter}

\item{scale}{scale parameter}
}
\value{
A vector of simulated values of length n.
}
\description{
Random Number Generation for Double Weibull Distribution
}
\keyword{internal}
