% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_models.R
\name{resid,tscmfit-method}
\alias{resid,tscmfit-method}
\title{Residual Method for tscmfit Class}
\usage{
\S4method{resid}{tscmfit}(object, trace = FALSE)
}
\arguments{
\item{object}{an object of class \linkS4class{tscmfit}.}

\item{trace}{extract trace instead of residuals.}
}
\description{
Residual Method for tscmfit Class
}
