% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{resid,tscopulafit-method}
\alias{resid,tscopulafit-method}
\title{Residual Method for tscopulafit Class}
\usage{
\S4method{resid}{tscopulafit}(object, trace = FALSE)
}
\arguments{
\item{object}{an object of class \linkS4class{tscopulafit}.}

\item{trace}{extract trace instead of residuals.}
}
\description{
Residual Method for tscopulafit Class
}
