% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacopula.R
\name{resid_armacopula}
\alias{resid_armacopula}
\title{Residual function for armacopula object}
\usage{
resid_armacopula(object, data = NA, trace = FALSE)
}
\arguments{
\item{object}{a fitted armacopula object.}

\item{data}{the data to which copula is fitted.}

\item{trace}{extract trace instead of residuals.}
}
\value{
vector of model residuals
}
\description{
Residual function for armacopula object
}
\keyword{internal}
