% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula.R
\name{resid_dvinecopula}
\alias{resid_dvinecopula}
\title{Residual function for dvinecopula object}
\usage{
resid_dvinecopula(object, data = NA, trace = FALSE)
}
\arguments{
\item{object}{a fitted dvinecopula object.}

\item{data}{the data to which copula is fitted.}

\item{trace}{extract trace instead of residuals.}
}
\value{
vector of model residuals
}
\description{
Residual function for dvinecopula object
}
\keyword{internal}
