% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{resid_dvinecopula2}
\alias{resid_dvinecopula2}
\title{Residual function for dvinecopula2 object}
\usage{
resid_dvinecopula2(object, data = NA, trace = FALSE)
}
\arguments{
\item{object}{a fitted dvinecopula2 object.}

\item{data}{the data to which copula is fitted.}

\item{trace}{extract trace instead of residuals.}
}
\value{
vector of model residuals
}
\description{
Residual function for dvinecopula2 object
}
\keyword{internal}
