% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula3.R
\name{resid_dvinecopula3}
\alias{resid_dvinecopula3}
\title{Residual function for dvinecopula3 object}
\usage{
resid_dvinecopula3(object, data = NA, trace = FALSE)
}
\arguments{
\item{object}{a fitted dvinecopula3 object.}

\item{data}{the data to which copula is fitted.}

\item{trace}{extract trace instead of residuals.}
}
\value{
vector of model residuals
}
\description{
Residual function for dvinecopula3 object
}
\keyword{internal}
