% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopulavt.R
\name{resid_dvinecopulavt}
\alias{resid_dvinecopulavt}
\title{Residual function for dvinecopulavt object}
\usage{
resid_dvinecopulavt(object, data = NA, trace = FALSE)
}
\arguments{
\item{object}{a fitted dvinecopulavt object.}

\item{data}{the data to which copula is fitted.}

\item{trace}{extract trace instead of residuals.}
}
\value{
vector of model residuals
}
\description{
Residual function for dvinecopulavt object
}
\keyword{internal}
