% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarmacopula.R
\name{resid_sarmacopula}
\alias{resid_sarmacopula}
\title{Residual function for sarmacopula object}
\usage{
resid_sarmacopula(object, data = NA, trace = FALSE)
}
\arguments{
\item{object}{a fitted sarmacopula object.}

\item{data}{the data to which copula is fitted.}

\item{trace}{extract trace instead of residuals.}
}
\value{
vector of model residuals
}
\description{
Residual function for sarmacopula object
}
\keyword{internal}
