% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdvinecopula.R
\name{resid_sdvinecopula}
\alias{resid_sdvinecopula}
\title{Residual function for sdvinecopula object}
\usage{
resid_sdvinecopula(object, data = NA, trace = FALSE)
}
\arguments{
\item{object}{a fitted sdvinecopula object.}

\item{data}{the data to which copula is fitted.}

\item{trace}{extract trace instead of residuals.}
}
\value{
vector of model residuals
}
\description{
Residual function for sdvinecopula object
}
\keyword{internal}
