% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{rlaplace}
\alias{rlaplace}
\title{Random Number Generation for Laplace Distribution}
\usage{
rlaplace(n, mu = 0, scale = 1)
}
\arguments{
\item{n}{size}

\item{mu}{location parameter}

\item{scale}{scale parameter}
}
\value{

}
\description{
Random Number Generation for Laplace Distribution
}
\keyword{internal}
