% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{rlgamma2}
\alias{rlgamma2}
\title{Random number generating for lgamma2}
\usage{
rlgamma2(n, sh1 = 1, r1 = 2, sh2 = 1, r2 = 2, p0 = 0)
}
\arguments{
\item{n}{size}

\item{sh1}{shapelog parameter of right branch}

\item{r1}{ratelog parameter of right branch}

\item{sh2}{shapelog parameter of left branch}

\item{r2}{ratelog parameter of left branch}

\item{p0}{mixing parameter}
}
\value{

}
\description{
Random number generating for lgamma2
}
\keyword{internal}
