% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{rNIG}
\alias{rNIG}
\title{Random Number Generation for NIG}
\usage{
rNIG(n, alphabar, gamma, mu, sigma)
}
\arguments{
\item{n}{size}

\item{sigma}{}
}
\value{

}
\description{
Random Number Generation for NIG
}
\keyword{internal}
