% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{rsdoubleweibull}
\alias{rsdoubleweibull}
\title{Random Number Generation for Skew Double Weibull Distribution}
\usage{
rsdoubleweibull(n, mu = 0, shape = 1, scale = 1, gamma = 1)
}
\arguments{
\item{n}{size}

\item{mu}{location parameter}

\item{shape}{shape parameter}

\item{scale}{scale parameter}

\item{gamma}{skewness parameter}
}
\value{
A vector of values of length n.
}
\description{
Random Number Generation for Skew Double Weibull Distribution
}
\keyword{internal}
