% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{rslaplace}
\alias{rslaplace}
\title{Random Number Generation for Skew Laplace Distribution}
\usage{
rslaplace(n, mu = 0, scale = 1, gamma = 1)
}
\arguments{
\item{n}{size}

\item{mu}{location parameter}

\item{scale}{scale parameter}

\item{gamma}{skewness paramater}
}
\value{

}
\description{
Random Number Generation for Skew Laplace Distribution
}
\keyword{internal}
