% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{rsst}
\alias{rsst}
\title{Random Number Generation for Skewed Student t}
\usage{
rsst(n, df, gamma, mu, sigma)
}
\arguments{
\item{n}{size}

\item{df}{degrees of freedom}

\item{gamma}{skewness parameter}

\item{mu}{location parameter}

\item{sigma}{scale parameter}
}
\value{

}
\description{
Random Number Generation for Skewed Student t
}
\keyword{internal}
