% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{rst}
\alias{rst}
\title{Random Number Generation for Student t}
\usage{
rst(n, df, mu, sigma)
}
\arguments{
\item{n}{size}

\item{df}{degrees of freedom}

\item{mu}{location parameter}

\item{sigma}{scale parameter}
}
\value{

}
\description{
Random Number Generation for Student t
}
\keyword{internal}
