% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{rweibull2}
\alias{rweibull2}
\title{Random number generating for weibull2}
\usage{
rweibull2(n, sh1 = 1, sc1 = 1, sh2 = 1, sc2 = 1, p0 = 0)
}
\arguments{
\item{n}{size}

\item{sh1}{shape parameter of right branch}

\item{sc1}{scale parameter of right branch}

\item{sh2}{shape parameter of left branch}

\item{sc2}{scale parameter of left branch}

\item{p0}{mixing parameter}
}
\value{

}
\description{
Random number generating for weibull2
}
\keyword{internal}
