% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{safe_ses}
\alias{safe_ses}
\title{Calculate Standard Errors Safely}
\usage{
safe_ses(hess)
}
\arguments{
\item{hess}{a Hessian matrix from a model fit.}
}
\value{
a vector of standard errors.
}
\description{
Calculate Standard Errors Safely
}
