% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarmacopula.R
\name{sarma2arma}
\alias{sarma2arma}
\title{Transform a sarmacopula object into an armacopula object}
\usage{
sarma2arma(object)
}
\arguments{
\item{object}{an object of class \linkS4class{sarmacopula}.}
}
\value{
An object of class \linkS4class{armacopula}.
}
\description{
Transform a sarmacopula object into an armacopula object
}
\examples{
sarma2arma(sarmacopula(list(ar = 0.5, ma = 0.4, sar = 0.2, sma = 0.6), period = 4))
}
