% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarmacopula.R
\name{sarma2dvine}
\alias{sarma2dvine}
\title{Transform a sarmacopula into a dvinecopula2 object}
\usage{
sarma2dvine(object)
}
\arguments{
\item{object}{an object of class \linkS4class{sarmacopula}.}
}
\value{
An object of class \linkS4class{dvinecopula2}.
}
\description{
Transform a sarmacopula into a dvinecopula2 object
}
\examples{
sarma2dvine(sarmacopula(list(ar = 0.5, ma = 0.4, sar = 0.2, sma = 0.6), period = 4))
}
