% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarmacopula.R
\name{sarmacopula}
\alias{sarmacopula}
\title{Constructor function for SARMA copula process}
\usage{
sarmacopula(pars = list(ar = 0, ma = 0, sar = 0, sma = 0), period = 4)
}
\arguments{
\item{pars}{list consisting of vector of AR parameters named `ar`
and vector of MA parameters named `ma`, SAR parameters named `sar`
and vector of SMA parameters named `sma`.}

\item{period}{period of seasonal model.}
}
\value{
An object of class \linkS4class{sarmacopula}.
}
\description{
Constructor function for SARMA copula process
}
\examples{
sarmacopula(list(ar = 0.5, ma = 0.4, sar = 0.2, sma = 0.6), period = 4)
}
