% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarmacopula.R
\name{sarmavec2list}
\alias{sarmavec2list}
\title{Turn vector of SARMA parameters into list}
\usage{
sarmavec2list(theta, order)
}
\arguments{
\item{theta}{vector of SARMA model parameters}

\item{order}{order of model}
}
\value{
a list containing SARMA parameters in components ar, ma, sar and sma
}
\description{
Turn vector of SARMA parameters into list
}
