% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{sdoubleweibull}
\alias{sdoubleweibull}
\alias{dsdoubleweibull}
\alias{psdoubleweibull}
\alias{qsdoubleweibull}
\alias{rsdoubleweibull}
\title{Skew double Weibull distribution}
\usage{
dsdoubleweibull(x, mu = 0.05, shape = 1, scale = 1, gamma = 1, log = FALSE)

psdoubleweibull(q, mu = 0.05, shape = 1, scale = 1, gamma = 1)

qsdoubleweibull(p, mu = 0.05, shape = 1, scale = 1, gamma = 1)

rsdoubleweibull(n, mu = 0.05, shape = 1, scale = 1, gamma = 1)
}
\arguments{
\item{x}{vector of values}

\item{mu}{location parameter}

\item{shape}{shape parameter}

\item{scale}{scale parameter}

\item{gamma}{skewness parameter}

\item{log}{flag for log density}

\item{q}{vector of quantiles}

\item{p}{vector of probabilities}

\item{n}{number of observations}
}
\description{
Skew double Weibull distribution
}
