% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdvinecopula.R
\docType{class}
\name{sdvinecopula-class}
\alias{sdvinecopula-class}
\alias{coef,sdvinecopula-method}
\alias{kendall,sdvinecopula-method}
\alias{show,sdvinecopula-method}
\alias{sim,sdvinecopula-method}
\alias{predict,sdvinecopula-method}
\title{Stationary d-vine copula processes}
\usage{
\S4method{coef}{sdvinecopula}(object)

\S4method{kendall}{sdvinecopula}(object, lagmax = 20)

\S4method{show}{sdvinecopula}(object)

\S4method{sim}{sdvinecopula}(object, n = 1000)

\S4method{predict}{sdvinecopula}(object, data, x, type = "df")
}
\arguments{
\item{object}{an object of the class.}

\item{lagmax}{maximum value of lag to be considered.}

\item{n}{length of realization.}

\item{data}{vector of past data values.}

\item{x}{vector of arguments of prediction function.}

\item{type}{type of prediction function ("df" for density, "qf" for quantile function
or "dens" for density).}
}
\description{
Class of objects for stationary d-vine copula processes. See \link{sdvinecopula} for more details.
}
\section{Methods (by generic)}{
\itemize{
\item \code{coef(sdvinecopula)}: Coef Method for sdvinecopula class

\item \code{kendall(sdvinecopula)}: Calculate Kendall's tau values for pair copulas in type 3 d-vine copula

\item \code{show(sdvinecopula)}: Show method for sdvinecopula class

\item \code{sim(sdvinecopula)}: Simulation method for sdvinecopula class

\item \code{predict(sdvinecopula)}: Prediction method for sdvinecopula class

}}
\section{Slots}{

\describe{
\item{\code{name}}{name of the stationary d-vine copula process.}

\item{\code{modelspec}}{list containing the family, rotation, and name of KPACF}

\item{\code{pars}}{list comprising of the parameters.}
}}

