% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdvinecopula.R
\name{sdvinecopula}
\alias{sdvinecopula}
\title{Constructor function for sdvinecopula process}
\usage{
sdvinecopula(
  pars = list(ar = 0, ma = 0),
  kpacf = "kpacf_arma",
  family = NULL,
  posrot = 0,
  negrot = 90,
  basefamily = "gauss",
  baseposrot = 0,
  basenegrot = 0,
  auxpar = NA,
  tautol = 1e-04,
  maxlag = Inf
)
}
\arguments{
\item{pars}{a list containing the parameters of the model}

\item{kpacf}{a character string giving the name of the Kendall pacf (default is kpacf_arma)}

\item{family}{vector of family names for copula substitutions}

\item{posrot}{vector of rotations for substituted families under positive dependence (default is 0)}

\item{negrot}{vector of rotations for substituted families under negative dependence (default is 90)}

\item{basefamily}{scalar specifying base copula family (default is "gauss")}

\item{baseposrot}{scalar specifying copula rotation under positive dependence (default is 0)}

\item{basenegrot}{scalar specifying copula rotation under negative dependence (default is 0)}

\item{auxpar}{vector of additional parameters for two-parameter copulas}

\item{tautol}{scalar value at which kpacf is truncated}

\item{maxlag}{a scalar which can be used to force a given value for maximum lag}
}
\value{
An object of class \linkS4class{sdvinecopula}.
}
\description{
This function sets up a stationary d-vine process of finite or infinite order. In general, models models consist of
base copulas and substituted copulas, which form a copula sequence. The copulas are parameterized using the
Kendall partial autocorrelation function (kpacf) specified by the \code{kpacf} argument.
}
\details{
The default choice is the
kpacf of a standard ARMA process which is implemented in the function \code{\link{kpacf_arma}}. The parameters
of the kpacf should be set as a named list using the \code{pars} argument; the required parameters should usually
be clear from the documentation of the chosen kpacf function and must be correctly named.

The base copula sequence is specified by the name of a single 1-parameter copula in \code{basefamily},
which can be Gauss, Frank, Gumbel, Clayton or Joe. For choices other than Gauss or Frank, the
user must specify the rotation that should be used for positive dependencies (0 or 180) and the rotation that
should be used for negative dependencies (90 or 270). These are specified by the arguments \code{baseposrot} and
\code{basenegrot} respectively.

The number of copulas to be replaced is determined by the length of the \code{family} vector, which should be
a vector of character strings representing copula names. The default value is \code{NULL} meaning no substitutions.
If \code{family} has length one, the same copula name is repeated for all substitutions.
The substituted families can be Gauss, Gumbel, Clayton, Joe, Frank, t and BB1 copulas as implemented by the
\code{\link[rvinecopulib]{bicop_dist}} in the \code{rvinecopulib} package.

For the substituted copulas (other than Gauss, t and Frank) the user should specify the rotation that should be used for
positive dependencies (0 or 180) and the rotation that should be used for negative dependencies (90 or 270). These are
specified by the arguments \code{posrot} and \code{negrot} respectively, which have default values 0 and 90.
These vectors should be the same length as \code{family} or length one, in which case the values are repeated.

In practice, the sequence of copulas will be truncated at the last copula for which the kpacf exceeds \code{tautol}.
The \code{maxlag} parameter is typically used to force the truncation to take place at a lower lag (to increase speed).
This can also be achieved by increasing the value of \code{tautol}.

If one or more of the substituted copulas are t or BB1 copulas the argument \code{auxpar} should be used to
specify the additional parameters. These are the degree-of-freedom parameter for t and the delta parameter for BB1;
the former must be greater or equal 2 and the latter greater or equal 1.
}
\examples{
sdvinecopula(pars = list(ar = 0.95, ma = 0.85), kpacf = "kpacf_arma",
family = c("Gumbel", "clayton"), posrot = c(0, 180), negrot = c(90, 270),
tautol = 1e-04)
}
