% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdvinecopula.R
\name{sdvinecopula_objective}
\alias{sdvinecopula_objective}
\title{Objective function for sdvinecopula process}
\usage{
sdvinecopula_objective(theta, modelspec, u)
}
\arguments{
\item{theta}{parameters of kpacf}

\item{modelspec}{list specifying model}

\item{u}{data}
}
\value{
Value of objective function at parameters.
}
\description{
Objective function for sdvinecopula process
}
\keyword{internal}
