% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{setoptions}
\alias{setoptions}
\title{Set Optional Choices for tscopula Fitting}
\usage{
setoptions(tsoptions, defaults)
}
\arguments{
\item{tsoptions}{a list of options chosen by user.}

\item{defaults}{a list of defaults specified in function.}
}
\value{
the definitive list of options to be used in function.
}
\description{
Set Optional Choices for tscopula Fitting
}
\keyword{internal}
