% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_vtscopula.R
\name{setwcopula}
\alias{setwcopula}
\title{Extract W-Copula}
\usage{
setwcopula(x)
}
\arguments{
\item{x}{an object of class \linkS4class{tscopula}.}
}
\value{
A description of the W-copula.
}
\description{
Extract W-Copula
}
\keyword{internal}
