% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacopula.R
\name{show,armacopula-method}
\alias{show,armacopula-method}
\title{Show Method for ARMA copula process}
\usage{
\S4method{show}{armacopula}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{armacopula}.}
}
\value{
summary of \linkS4class{armacopula}.
}
\description{
Show Method for ARMA copula process
}
