% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{show,dvinecopula2-method}
\alias{show,dvinecopula2-method}
\title{Show Method for dvinecopula2 Class}
\usage{
\S4method{show}{dvinecopula2}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{dvinecopula2}.}
}
\value{

}
\description{
Show Method for dvinecopula2 Class
}
