% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{show,margin-method}
\alias{show,margin-method}
\title{Show Method for margin Class}
\usage{
\S4method{show}{margin}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{margin}.}
}
\value{
A summary of an object of class \linkS4class{margin}.
}
\description{
Show Method for margin Class
}
