% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{show,tscopulafit-method}
\alias{show,tscopulafit-method}
\title{Show method for tscopulafit objects}
\usage{
\S4method{show}{tscopulafit}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{tscopulafit}.}
}
\value{
parameters of tscopulafit model
}
\description{
Show method for tscopulafit objects
}
