% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_vtscopula.R
\name{show,vtscopula-method}
\alias{show,vtscopula-method}
\title{Show Method for vtscopula objects}
\usage{
\S4method{show}{vtscopula}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{vtscopula}.}
}
\value{
Summary of object of class \linkS4class{vtscopula}.
}
\description{
Show Method for vtscopula objects
}
