% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacopula.R
\name{sigmastarma}
\alias{sigmastarma}
\title{Standard Deviation of Innovations for armacopula}
\usage{
sigmastarma(x)
}
\arguments{
\item{x}{an object of class \linkS4class{armacopula}.}
}
\value{
The standard deviation of the standardized ARMA innovation distribution.
}
\description{
Uses the function \code{\link[ltsa]{tacvfARMA}} in the ltsa library.
}
\examples{
sigmastarma(armacopula(list(ar = c(0.5, 0.4), ma = -0.8)))
}
