% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacopula.R
\name{sim,armacopula-method}
\alias{sim,armacopula-method}
\title{Simulation Method for armacopula Class}
\usage{
\S4method{sim}{armacopula}(x, n = 1000)
}
\arguments{
\item{x}{an object of class \linkS4class{armacopula}.}

\item{n}{length of realization.}
}
\value{
A realization of a time series copula process.
}
\description{
Simulation Method for armacopula Class
}
\examples{
sim(armacopula(list(ar = c(0.5, 0.4), ma = -0.8)), n = 1000)
}
