% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{sim,margin-method}
\alias{sim,margin-method}
\title{Simulation Method for margin Class}
\usage{
\S4method{sim}{margin}(x, n = 1000)
}
\arguments{
\item{x}{an object of class \linkS4class{margin}.}

\item{n}{length of realization.}
}
\value{

}
\description{
Simulation Method for margin Class
}
\examples{
margmod <- margin("norm", pars = c(mean = 0, sd = 1))
sim(margmod, n = 500)
}
