% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_objects.R
\name{sim,swncopula-method}
\alias{sim,swncopula-method}
\title{Simulation Method for Strict White Noise Copula}
\usage{
\S4method{sim}{swncopula}(x, n = 1000)
}
\arguments{
\item{n}{numeric value for length of simulated realisation.}

\item{swncopula}{an object of class \linkS4class{swncopula}.}
}
\value{
a realisation of strict white noise of length n
}
\description{
Simulation Method for Strict White Noise Copula
}
\examples{
sim(swncopula())
}
