% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_models.R
\name{sim,tscm-method}
\alias{sim,tscm-method}
\title{Simulation Method for tscm class}
\usage{
\S4method{sim}{tscm}(x, n = 1000)
}
\arguments{
\item{x}{an object of class \linkS4class{tscm}.}

\item{n}{length of realization.}
}
\value{
A realization of the time series of length n
}
\description{
Simulation Method for tscm class
}
\examples{
mod <- tscm(dvinecopula(family = "gauss", pars = 0.5), margin("weibull2"))
sim(mod)
}
