% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{sim,tscopulafit-method}
\alias{sim,tscopulafit-method}
\title{Simulation Method for tscopulafit Class}
\usage{
\S4method{sim}{tscopulafit}(x, n = 1000)
}
\arguments{
\item{x}{an object of class \linkS4class{tscopulafit}.}

\item{n}{length of realization.}
}
\value{
A realization of a time series copula process.
}
\description{
Simulation Method for tscopulafit Class
}
\examples{
ar1 <- armacopula(list(ar = 0.7))
data <- sim(ar1, 1000)
ar1fit <- fit(ar1, data)
sim(ar1fit)
}
