% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_vtscopula.R
\name{sim,vtscopula-method}
\alias{sim,vtscopula-method}
\title{Simulation Method for vtscopula Class}
\usage{
\S4method{sim}{vtscopula}(x, n = 1000)
}
\arguments{
\item{x}{an object of class \linkS4class{vtscopula}.}

\item{n}{length of realization.}
}
\value{
A realization of a time series copula process.
}
\description{
Simulation Method for vtscopula Class
}
\examples{
copobject <- armacopula(pars = list(ar = 0.6, ma = 0.2))
sim(vtscopula(copobject, Vtransform = V2p()))
}
