% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_objects.R
\name{sim}
\alias{sim}
\title{Generic for simulating time series copula models}
\usage{
sim(object, ...)
}
\arguments{
\item{object}{an object of the model class.}

\item{...}{further arguments to be passed to the simulation.}
}
\value{
A simulated realization from the time series model.
}
\description{
Methods are available for objects of class \linkS4class{swncopula},
\linkS4class{armacopula},
\linkS4class{dvinecopula}, \linkS4class{dvinecopula2},
\linkS4class{margin} and \linkS4class{tscm}.
}
