% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{sst}
\alias{sst}
\alias{psst}
\alias{qsst}
\alias{dsst}
\alias{rsst}
\title{Skew Student t distribution}
\usage{
psst(q, df = 10, gamma = 1, mu = 0, sigma = 1)

qsst(p, df, gamma, mu, sigma)

dsst(x, df, gamma, mu, sigma, log = FALSE)

rsst(n, df, gamma, mu, sigma)
}
\arguments{
\item{q}{vector of quantiles.}

\item{df}{degrees of freedom.}

\item{gamma}{skewness parameter.}

\item{mu}{location parameter.}

\item{sigma}{scale parameter.}

\item{p}{vector of probabilities.}

\item{x}{vector of values.}

\item{log}{flag for log density.}

\item{n}{number of observations.}
}
\value{
A vector of density, distribution function, quantile or random values.
}
\description{
Skew Student t distribution
}
