% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{st}
\alias{st}
\alias{pst}
\alias{qst}
\alias{dst}
\alias{rst}
\title{Student t distribution}
\usage{
pst(q, df = 10, mu = 0, sigma = 1)

qst(p, df, mu, sigma)

dst(x, df, mu, sigma, log = FALSE)

rst(n, df, mu, sigma)
}
\arguments{
\item{q}{vector of quantiles}

\item{df}{degrees of freedom}

\item{mu}{location parameter}

\item{sigma}{scale parameter}

\item{p}{vector of probabilities}

\item{x}{vector of values}

\item{log}{flag for log density}

\item{n}{number of observations}
}
\description{
Student t distribution
}
