% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{st0}
\alias{st0}
\alias{pst0}
\alias{qst0}
\alias{dst0}
\alias{rst0}
\title{Centred Student t distribution}
\usage{
pst0(q, df = 10, sigma = 1)

qst0(p, df, sigma)

dst0(x, df, sigma, log = FALSE)

rst0(n, df, sigma)
}
\arguments{
\item{q}{vector of quantiles.}

\item{df}{degrees of freedom.}

\item{sigma}{scale parameter.}

\item{p}{vector of probabilities.}

\item{x}{vector of values.}

\item{log}{flag for log density.}

\item{n}{number of observations.}
}
\value{
A vector of density, distribution function, quantile or random values.
}
\description{
Centred Student t distribution
}
