% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacopula.R
\name{starmaStateSpace}
\alias{starmaStateSpace}
\title{State Space Representation for standardized ARMA model}
\usage{
starmaStateSpace(ar, ma, order)
}
\arguments{
\item{order}{}
}
\value{

}
\description{
State Space Representation for standardized ARMA model
}
\keyword{internal}
