% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{strank}
\alias{strank}
\title{Calculate standardized ranks of data}
\usage{
strank(x)
}
\arguments{
\item{x}{a vector or time series of data.}
}
\value{
A vector or time series of standardized ranks in the interval (0,1)
}
\description{
Calculate standardized ranks of data
}
\examples{
strank(rnorm(100))
}
