% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_objects.R
\name{swncopula}
\alias{swncopula}
\title{Constructor function for strict white noise copula process}
\usage{
swncopula()
}
\value{
Object of class \linkS4class{swncopula}.
}
\description{
Constructor function for strict white noise copula process
}
\examples{
swncopula()
}
