% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_models.R
\name{tsc_objectivea}
\alias{tsc_objectivea}
\title{Objective Function for Full Fit}
\usage{
tsc_objectivea(theta, modelspec, modeltype, dens, cdf, margfixed, y)
}
\arguments{
\item{avoidzero}{}
}
\value{

}
\description{
Objective Function for Full Fit
}
\keyword{internal}
