% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_models.R
\name{tscm}
\alias{tscm}
\title{Constructor function for time series}
\usage{
tscm(tscopula, margin = new("margin", name = "unif"))
}
\arguments{
\item{tscopula}{an object of class \linkS4class{tscopula}.}

\item{margin}{an object of class \linkS4class{margin}.}
}
\value{
An object of class \linkS4class{tscm}.
}
\description{
Constructor function for time series
}
\examples{
tscm(dvinecopula(family = "gauss", pars = 0.5), margin("doubleweibull"))
}
