\name{tscopula-package}
\alias{tscopula-package}
\alias{tscopula}
\docType{package}
\title{
  Time Series Copulas
}
\description{
  This package contains functions for the analysis of time series using copula models. A full description of a strictly stationary time series can be obtained by choosing a univariate marginal distribution and a time series copula process, i.e. a serially dependent process of uniform random variables. Examples of the latter are the copula processes of Gaussian ARMA models and D-vine copula processes. Methods are provided for simulation and estimation of time series copula models.
}
\author{
Alexander J. McNeil and Martin Bladt.

Maintainer: Alexander J. McNeil <alexanderjmcneil@gmail.com>
}
\references{
  McNeil, A.J. (2021). Modelling volatile time series with v-transforms and copula. Risks, 9(14).

  Bladt, M., & McNeil, A.J. (2020). Time series copula models using d-vines and v-transforms: an alternative to GARCH modelling. arXiv:2006.11088.
}
\keyword{ tscopula }
