% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopulavt.R
\name{tscopula_bicop_dist}
\alias{tscopula_bicop_dist}
\title{Helper function for building lists of objects to describe copula sequence}
\usage{
tscopula_bicop_dist(family, rotation, parameters)
}
\arguments{
\item{family}{name of copula family}

\item{rotation}{rotation for copula family}

\item{parameters}{parameters of copula}
}
\value{
a list containing specification of copula family
}
\description{
Helper function for building lists of objects to describe copula sequence
}
\keyword{internal}
