% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopulavt.R
\name{tscopula_dbicop}
\alias{tscopula_dbicop}
\title{Compute density of bivariate copula}
\usage{
tscopula_dbicop(u, family)
}
\arguments{
\item{u}{matrix of data with two columns}

\item{family}{list describing copula family}
}
\value{
a vector containing values of density
}
\description{
Compute density of bivariate copula
}
\keyword{internal}
