% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopulavt.R
\name{tscopula_hbicop}
\alias{tscopula_hbicop}
\title{Compute h function of bivariate copula}
\usage{
tscopula_hbicop(u, cond_var, family, inverse = FALSE)
}
\arguments{
\item{u}{matrix of data with two columns}

\item{cond_var}{conditioning variable (1 or 2)}

\item{family}{list describing copula family}

\item{inverse}{logical variable specifying whether inverse should be returned}
}
\value{
a vector containing values of density
}
\description{
Compute h function of bivariate copula
}
\keyword{internal}
