% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\docType{class}
\name{tscopulafit-class}
\alias{tscopulafit-class}
\title{Fitted Time Series Copula Processes}
\value{

}
\description{
Class of objects for time series copula processes.
}
\section{Slots}{

\describe{
\item{\code{name}}{name of time series copula process.}
}}

