% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\docType{class}
\name{tscopulaU-class}
\alias{tscopulaU-class}
\title{Time series copulas of class tscopulaU}
\description{
S4 Class union for basic time series copula types. These are
\linkS4class{armacopula}, \linkS4class{dvinecopula} and \linkS4class{dvinecopula2},
}
