% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{tsrelist}
\alias{tsrelist}
\title{Relist Function for tscopula Objects}
\usage{
tsrelist(ests, fulcrum = NA)
}
\arguments{
\item{ests}{a relistable vector of parameter estimates}

\item{fulcrum}{numerical value for fixed fulcrum or NA}
}
\value{
list of parameters for \linkS4class{tscopulafit} object.
}
\description{
Relist Function for tscopula Objects
}
\keyword{internal}
