% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{tsunlist}
\alias{tsunlist}
\title{Unlist Function for tscopula Objects}
\usage{
tsunlist(pars, fulcrum = NA)
}
\arguments{
\item{pars}{list of parameters}

\item{fulcrum}{numerical value for fixed fulcrum or NA}
}
\value{
relistable vector of parameters
}
\description{
Unlist Function for tscopula Objects
}
\keyword{internal}
