% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtransforms.R
\name{V2b}
\alias{V2b}
\title{Constructor function for 2-parameter beta v-transform}
\usage{
V2b(delta = 0.5, kappa = 1)
}
\arguments{
\item{delta}{a value in (0, 1) specifying the fulcrum of the v-transform.}

\item{kappa}{additional positive parameter of v-transform.}
}
\value{
An object of class \linkS4class{Vtransform}.
}
\description{
Constructor function for 2-parameter beta v-transform
}
\examples{
V2b(delta = 0.45, kappa = 1.2)
}
