% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtransforms.R
\name{V2p}
\alias{V2p}
\title{Constructor Function for 2-Parameter V-Transform}
\usage{
V2p(delta = 0.5, kappa = 1)
}
\arguments{
\item{delta}{a value in (0, 1) specifying the fulcrum of the v-transform.}

\item{kappa}{additional positive parameter of v-transform.}
}
\value{
A \code{Vtransform} object.
}
\description{
Constructor Function for 2-Parameter V-Transform
}
\examples{
V2p(delta = 0.45, kappa = 1.2)
}
