% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtransforms.R
\name{V3b}
\alias{V3b}
\title{Constructor Function for 3-Parameter Beta V-Transform}
\usage{
V3b(delta = 0.5, kappa = 1, xi = 1)
}
\arguments{
\item{delta}{a value in (0, 1) specifying the fulcrum of the v-transform.}

\item{kappa}{additional positive parameter of v-transform.}

\item{xi}{additional positive parameter of v-transform.}
}
\value{
A \code{Vtransform} object.
}
\description{
Constructor Function for 3-Parameter Beta V-Transform
}
\examples{
V3b(delta = 0.45, kappa = 1.2, xi = 1.2)
}
