% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtransforms.R
\name{V3p}
\alias{V3p}
\title{Constructor Function for 3-Parameter V-Transform}
\usage{
V3p(delta = 0.5, kappa = 1, xi = 1)
}
\arguments{
\item{delta}{a value in (0, 1) specifying the fulcrum of the v-transform.}

\item{kappa}{additional positive parameter of v-transform.}

\item{xi}{additional positive parameter of v-transform.}
}
\value{
A \code{Vtransform} object.
}
\description{
Constructor Function for 3-Parameter V-Transform
}
\examples{
V3p(delta = 0.45, kappa = 0.8, xi = 1.1)
}
