% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtransforms.R
\name{Vdegenerate}
\alias{Vdegenerate}
\title{Constructor Function for Degenerate V-Transform}
\usage{
Vdegenerate()
}
\value{
A \code{VtransformI} object.
}
\description{
Constructor Function for Degenerate V-Transform
}
\examples{
Vdegenerate()
}
