% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtransforms.R
\name{vdownprob}
\alias{vdownprob}
\title{Calculate Conditional Down Probability of V-Transform}
\usage{
vdownprob(x, v)
}
\arguments{
\item{x}{a \code{Vtransform} object.}

\item{v}{a vector or time series with values in [0, 1].}
}
\value{
A vector or time series of values of gradient.
}
\description{
Calculate Conditional Down Probability of V-Transform
}
\examples{
vdownprob(V2p(delta = 0.55, kapp = 1.2), c(0, 0.25, 0.5, 0.75, 1))
}
