% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtransforms.R
\name{Vlinear}
\alias{Vlinear}
\title{Constructor function for linear v-transform}
\usage{
Vlinear(delta = 0.5)
}
\arguments{
\item{delta}{a value in (0, 1) specifying the fulcrum of the v-transform.}
}
\value{
An o bject of class \linkS4class{VtransformI}.
}
\description{
Constructor function for linear v-transform
}
\examples{
Vlinear(delta = 0.45)
}
