% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtransforms.R
\name{Vsymmetric}
\alias{Vsymmetric}
\title{Constructor function for symmetric v-transform}
\usage{
Vsymmetric()
}
\value{
An object of class \linkS4class{VtransformI}.
}
\description{
Constructor function for symmetric v-transform
}
\examples{
Vsymmetric()
}
