% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_vtscopula.R
\name{vtparlist}
\alias{vtparlist}
\title{Extract parameters of vtscopula}
\usage{
vtparlist(object)
}
\arguments{
\item{object}{an object of class \linkS4class{vtscopula}.}
}
\value{
A list of parameters.
}
\description{
Extract parameters of vtscopula
}
\keyword{internal}
