% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtransforms.R
\docType{class}
\name{Vtransform-class}
\alias{Vtransform-class}
\alias{show,Vtransform-method}
\alias{coef,Vtransform-method}
\title{Class of v-transforms}
\usage{
\S4method{show}{Vtransform}(object)

\S4method{coef}{Vtransform}(object)
}
\arguments{
\item{object}{an object of the class.}
}
\description{
This is the class of v-transforms. It contains the \linkS4class{VtransformI} subclass consisting of v-transforms
with an analytical expression for the inverse.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Show method for Vtransform class

\item \code{coef}: Coef method for Vtransform class
}}

\section{Slots}{

\describe{
\item{\code{name}}{a name for the v-transform of class character.}

\item{\code{Vtrans}}{function to evaluate the v-transform.}

\item{\code{pars}}{vector containing the named parameters of the v-transform.}

\item{\code{gradient}}{function to evaluate the gradient of the v-transform.}
}}

\examples{
V2p(delta = 0.5, kappa = 1.2)
}
