% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_vtscopula.R
\docType{class}
\name{vtscopula-class}
\alias{vtscopula-class}
\title{VT Time Series Copula Processes}
\description{
Class of objects for V-transformed time series copula processes.
}
\section{Slots}{

\describe{
\item{\code{Vcopula}}{object of class \linkS4class{tscopulaU}.}

\item{\code{Vtransform}}{object of class \linkS4class{Vtransform}.}

\item{\code{Wcopula}}{object of class \linkS4class{tscopula}.}
}}

