% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_vtscopula.R
\name{vtscopula}
\alias{vtscopula}
\title{Constructor Function for vtscopula Object}
\usage{
vtscopula(tscopulaU, Vtransform = Vlinear(), Wcopula = swncopula())
}
\arguments{
\item{tscopulaU}{an object of class
\linkS4class{armacopula} or \linkS4class{dvinecopula}.}

\item{Vtransform}{an object of class \linkS4class{Vtransform}.}

\item{Wcopula}{an object of class \linkS4class{tscopula}.}
}
\value{
An object of class \linkS4class{vtscopula}.
}
\description{
Constructor Function for vtscopula Object
}
\examples{
copobject <- armacopula(pars = list(ar = 0.6, ma = 0.2))
vtscopula(copobject, Vtransform = V2p())
}
