% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_vtscopula.R
\name{wobjective}
\alias{wobjective}
\title{Additional objective for generalized processes}
\usage{
wobjective(theta, paircop, theta_vt, vtgrad, u)
}
\arguments{
\item{theta}{parameter of W copula}

\item{paircop}{specification of W copula}

\item{theta_vt}{parameter(s) of v-transform}

\item{vtgrad}{gradient of v-transform}

\item{u}{vector of data}
}
\value{
Value of additional objective for W-copula.
}
\description{
Additional objective for generalized processes
}
\keyword{internal}
