% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{acf2pacf}
\alias{acf2pacf}
\title{Compute partial autocorrelations from autocorrelations}
\usage{
acf2pacf(rho)
}
\arguments{
\item{rho}{vector of autocorrelation values (excluding 1)}
}
\value{
vector of partial autocorrelation values
}
\description{
Compute partial autocorrelations from autocorrelations
}
\examples{
rho <- ARMAacf(ar = -0.9, ma = 0.8, lag.max = 50)[-1]
alpha <- acf2pacf(rho)
}
