% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacopula.R
\docType{class}
\name{armacopula-class}
\alias{armacopula-class}
\title{ARMA Copula Processes}
\value{

}
\description{
Class of objects for ARMA copula processes.
}
\section{Slots}{

\describe{
\item{\code{name}}{name of ARMA copula process.}

\item{\code{modelspec}}{vector containing number of AR and MA parameters.}

\item{\code{pars}}{list comprising of the parameters.}
}}

