% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacopula.R
\name{armacopula}
\alias{armacopula}
\title{Constructor Function for ARMA copula process}
\usage{
armacopula(pars = list(ar = 0, ma = 0))
}
\arguments{
\item{pars}{a list of length two containing AR and MA parameters.}
}
\value{
An object of class \linkS4class{armacopula}.
}
\description{
Constructor Function for ARMA copula process
}
\examples{
armacopula(list(ar = 0.5, ma = 0.4))
}
