% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacopula.R
\name{armacopula_objective}
\alias{armacopula_objective}
\title{Objective Function for ARMA copula process}
\usage{
armacopula_objective(theta, modelspec, u)
}
\arguments{
\item{theta}{vector of parameters of ARMA process}

\item{modelspec}{vector containing model order (p,q)}

\item{u}{vector of data}
}
\value{
Value of objective function at parameters.
}
\description{
Objective Function for ARMA copula process
}
\keyword{internal}
