% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_models.R
\name{coerce,tscopula,tscm-method}
\alias{coerce,tscopula,tscm-method}
\title{Convert tscopula Object to tscm Object}
\usage{
\S4method{coerce}{tscopula,tscm}(from, to = "tsc", strict = TRUE)
}
\arguments{
\item{from}{a \linkS4class{tscopula} object.}

\item{to}{a \linkS4class{tscm} object.}
}
\value{
a \linkS4class{tscm} object.
}
\description{
Convert tscopula Object to tscm Object
}
