% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{dvinecopula2}
\alias{dvinecopula2}
\title{Constructor function for dvinecopula2 process}
\usage{
dvinecopula2(
  family = "gauss",
  rotation = 0,
  kpacf = "kpacf_arma",
  pars = list(ar = 0.1, ma = 0.1),
  maxlag = Inf,
  negtau = "none"
)
}
\arguments{
\item{family}{family name}

\item{rotation}{rotation}

\item{kpacf}{character string giving name of Kendal pacf}

\item{pars}{a list containing the parameters of each lag}

\item{maxlag}{scalar specifying maximum lag}

\item{negtau}{character specifiying treatment of negative tau values}
}
\value{
An object of class \linkS4class{dvinecopula2}.
}
\description{
Constructor function for dvinecopula2 process
}
\examples{
dvinecopula2(family = "joe", kpacf = "kpacf_arma",
pars = list(ar = 0.95, ma = -0.85), maxlag = 30)
}
