% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula.R
\name{dvinecopula_objective}
\alias{dvinecopula_objective}
\title{Objective Function for dvinecopula process}
\usage{
dvinecopula_objective(theta, modelspec, u)
}
\arguments{
\item{theta}{parameters of copulas}

\item{modelspec}{list of families of copulas}

\item{udata}{data}
}
\value{
Value of objective function at parameters.
}
\description{
Objective Function for dvinecopula process
}
\keyword{internal}
