% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopulavt.R
\docType{class}
\name{dvinecopulavt-class}
\alias{dvinecopulavt-class}
\alias{coef,dvinecopulavt-method}
\alias{show,dvinecopulavt-method}
\title{D-vine copula processes with v-transforms}
\usage{
\S4method{coef}{dvinecopulavt}(object)

\S4method{show}{dvinecopulavt}(object)
}
\arguments{
\item{object}{an object of the class.}
}
\description{
Class of objects for d-vine copula processes. See \link{dvinecopulavt} for more details.
}
\section{Methods (by generic)}{
\itemize{
\item \code{coef(dvinecopulavt)}: Coef Method for dvinecopulavt class

\item \code{show(dvinecopulavt)}: Show method for dvinecopulavt class

}}
\section{Slots}{

\describe{
\item{\code{name}}{name of the d-vine copula process.}

\item{\code{modelspec}}{list containing the family, rotation, and name of KPACF}

\item{\code{pars}}{list comprising of the parameters.}
}}

