% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopulavt.R
\name{dvinecopulavt}
\alias{dvinecopulavt}
\title{Constructor function for dvinecopulavt process}
\usage{
dvinecopulavt(
  family = "joe",
  rotation = 0,
  kpacf = "kpacf_arma",
  pars = list(ar = 0.1, ma = 0.1),
  maxlag = Inf
)
}
\arguments{
\item{family}{family name}

\item{rotation}{a scalar specifying the rotation (default is 0)}

\item{kpacf}{a character string giving the name of the Kendall pacf}

\item{pars}{a list containing the parameters of the model}

\item{maxlag}{a scalar specifying the maximum lag}
}
\value{
An object of class \linkS4class{dvinecopulavt}.
}
\description{
This function sets up a stationary d-vine process of finite or infinite order based on a single
inverse-v-transformed copula family from a subset of those that can be implemented using
\code{\link[rvinecopulib]{bicop_dist}} in the \code{rvinecopulib} package.
}
\details{
The permitted choices of base copula family are currently Joe, Gumbel or Clayton survival. Thus,
if Clayton is chosen, the \code{rotation} argument must be set to 180, while if Joe or Gumbel are chosen, the
\code{rotation} argument must be zero (which is the default); any other options will return an error

The copulas are parameterized using the Kendall partial autocorrelation function (kpacf)
of the base copula sequence specified
by the \code{kpacf} argument. The default choice is the kpacf of a standard ARMA process which is
implemented in the function \code{\link{kpacf_arma}}. The parameters
of the kpacf should be set as a list using the \code{pars} argument; the required parameters should usually
be clear from the documentation of the chosen kpacf function and must be correctly named.

For data showing stochastic volatility, we expect positive serial dependencies in the base copula sequence.
For this reason, we do not consider models where the kpacf takes negative values.

The \code{maxlag} parameter specifies the maximum lag of the process; a finite number gives a finite-order
stationary d-vine process, but the parameter may also be set to \code{Inf} for an infinite-order process.
}
\examples{
dvinecopulavt(family = "joe", kpacf = "kpacf_arma",
pars = list(ar = 0.95, ma = -0.85), maxlag = 30)
}
