% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{fit}
\alias{fit}
\title{New Generic for Estimating Time Series Models}
\usage{
fit(x, y, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{y}{a vector or time series of data.}

\item{...}{further arguments to be passed on.}
}
\value{
An object of the fitted model class.
}
\description{
Methods are available for objects of class \linkS4class{tscopula},
\linkS4class{vtscopula}, \linkS4class{margin}
\linkS4class{tsc} and \linkS4class{gvtcopar}.
}
