% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_models.R
\name{fitFULLa}
\alias{fitFULLa}
\title{Fit tscm jointly}
\usage{
fitFULLa(x, y, tsoptions, control)
}
\arguments{
\item{x}{an object of class \linkS4class{tscm}.}

\item{y}{a vector or time series of data.}

\item{tsoptions}{list of variables}

\item{control}{list of control parameters to be passed to the
\code{\link[stats]{optim}} function.}
}
\value{
An object of class \linkS4class{tscmfit}.
}
\description{
Fit tscm jointly
}
\keyword{internal}
