% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{glag}
\alias{glag}
\title{Generalized lagging function}
\usage{
glag(x, lagmax = 20, glagplot = FALSE)
}
\arguments{
\item{x}{an object of class \linkS4class{tscopulafit}.}

\item{lagmax}{maximum value for lag.}

\item{glagplot}{logical value indicating generalized lag plot.}
}
\description{
Generalized lagging function
}
