% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_objects.R
\name{glag}
\alias{glag}
\title{New Generic for Generalized Lagging of Data}
\usage{
glag(x, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{...}{further arguments to be passed on.}
}
\value{
A realization from the time series model.
}
\description{
Methods are available for objects of class \linkS4class{tscopulafit}
 and \linkS4class{tscmfit}.
}
