% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula.R
\name{glag_for_dvinecopula}
\alias{glag_for_dvinecopula}
\title{Generalized lagging for fitted dvinecopula objects}
\usage{
glag_for_dvinecopula(copula, data, lagmax, glagplot = FALSE)
}
\arguments{
\item{copula}{a dvinecopula object}

\item{data}{the data to which copula is fitted}

\item{lagmax}{the maximum lag value.}

\item{glagplot}{logical value indicating generalized lag plot.}
}
\description{
Generalized lagging for fitted dvinecopula objects
}
\keyword{internal}
