% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopulavt.R
\name{hbicop}
\alias{hbicop}
\title{h-function for linear inverse-v-transformed copula}
\usage{
hbicop(u, cond_var, family, delta, inverse = FALSE)
}
\arguments{
\item{u}{two-column matrix of data at which h-function is evaluated}

\item{cond_var}{identity of conditioning variable (1/2)}

\item{family}{name of copula family}

\item{delta}{vector of two fulcrum values}

\item{inverse}{logical variable specifying whether inverse is taken}
}
\value{
vector of values of h-function
}
\description{
h-function for linear inverse-v-transformed copula
}
\keyword{internal}
