% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula.R
\name{IRblatt}
\alias{IRblatt}
\title{Calculate inverse Rosenblatt function}
\usage{
IRblatt(pc_list, data, q)
}
\arguments{
\item{pc_list}{a list of pair copulas.}

\item{data}{vector of past data values.}

\item{q}{vector of arguments of inverse Rosenblatt function.}
}
\value{
a vector of values with same length as \code{q}.
}
\description{
Calculate inverse Rosenblatt function
}
\keyword{internal}
