% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_objects.R
\name{kendall}
\alias{kendall}
\title{Generic for Kendall correlations}
\usage{
kendall(object, ...)
}
\arguments{
\item{object}{an object of the model class.}

\item{...}{further arguments to be passed to Kendall calculation.}
}
\value{
A vector of Kendall correlations.
}
\description{
Methods are available for objects of class \linkS4class{armacopula},
\linkS4class{dvinecopula}, \linkS4class{dvinecopula2} and \linkS4class{vtscopula}.
}
