% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{kpacf_arma}
\alias{kpacf_arma}
\title{KPACF of ARMA Process}
\usage{
kpacf_arma(k, theta)
}
\arguments{
\item{k}{vector of lags}

\item{theta}{list with components ar and ma specifying the ARMA parameters}
}
\value{
vector of Kendall partial autocorrelations for each lag k
}
\description{
KPACF of ARMA Process
}
