% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula2.R
\name{ktau_to_par}
\alias{ktau_to_par}
\title{Transform Kendall's tau values to copula parameters}
\usage{
ktau_to_par(family, tau)
}
\arguments{
\item{family}{name of copula family}

\item{tau}{value of Kendall's tau}
}
\value{
Scalar value of parameter giving tau for a copula family.
}
\description{
Transform Kendall's tau values to copula parameters
}
\keyword{internal}
