% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\docType{class}
\name{margin-class}
\alias{margin-class}
\title{Marginal Model for Time Series}
\value{

}
\description{
Class of objects for marginal models for stationary time series. The
object is given a name and there must exist functions pname, qname,
dname and rname. For example, the object could be named norm and
make use of \code{\link[stats]{pnorm}}, \code{\link[stats]{qnorm}},
\code{\link[stats]{dnorm}} and \code{\link[stats]{rnorm}}.
As well as the parameters of the distribution, dname must have the
logical argument log specifying whether log density should be computed.
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the marginal model class.}

\item{\code{pars}}{a numeric vector containing the named parameters of the distribution
which are passed as arguments to pname, qname, dname and rname.}
}}

\examples{
new("margin", name = "norm", pars = c(mu = 0, sigma = 1))
}
