% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\docType{class}
\name{marginfit-class}
\alias{marginfit-class}
\title{Fitted Marginal Model for Time Series}
\description{
Fitted Marginal Model for Time Series
}
\section{Slots}{

\describe{
\item{\code{margin}}{an object of class \linkS4class{margin}.}

\item{\code{data}}{numeric vector or time series of data.}

\item{\code{fit}}{a list containing details of the maximum likelihood fit.}
}}

