% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula3.R
\name{mklist_dvine3}
\alias{mklist_dvine3}
\title{Make list of pair copulas for dvinecopula3 object}
\usage{
mklist_dvine3(x, maxlag, truncate, tol = 1)
}
\arguments{
\item{x}{an object of class dvinecopula3}

\item{maxlag}{maximum possible lag to consider}

\item{truncate}{logical variable stating whether to truncate copulas with negligible dependence}

\item{tol}{tolerance for truncation}
}
\value{
a list of pair copulas
}
\description{
Make list of pair copulas for dvinecopula3 object
}
\keyword{internal}
