% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_margins.R
\name{NIG}
\alias{NIG}
\alias{pNIG}
\alias{qNIG}
\alias{dNIG}
\alias{rNIG}
\title{Normal inverse Gaussian distribution}
\usage{
pNIG(q, shape = 1, gamma = 0, mu = 0, sigma = 1)

qNIG(p, shape = 1, gamma = 0, mu = 0, sigma = 1)

dNIG(x, shape = 1, gamma = 0, mu = 0, sigma = 1, log = FALSE)

rNIG(n, shape = 1, gamma = 0, mu = 0, sigma = 0)
}
\arguments{
\item{q}{vector of quantiles}

\item{shape}{shape parameter}

\item{gamma}{skewness parameter}

\item{mu}{location parameter}

\item{sigma}{scale parameter}

\item{p}{vector of probabilities}

\item{x}{vector of values}

\item{n}{number of observations}
}
\description{
Normal inverse Gaussian distribution
}
