% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_vtarma.R
\name{pcondvtarma}
\alias{pcondvtarma}
\title{Conditional distribution function of VT-ARMA Process}
\usage{
pcondvtarma(q, tscmmod, armamean, armasd)
}
\arguments{
\item{q}{point at which CDF should be calculated.}

\item{tscmmod}{an object of class \linkS4class{tscmfit} based on underlying copula
of class \linkS4class{armacopula}.}

\item{armamean}{conditional mean of underlying ARMA process.}

\item{armasd}{conditional standard deviation of underlying ARMA process.}
}
\value{
a scalar value.
}
\description{
Conditional distribution function of VT-ARMA Process
}
\keyword{internal}
