% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_models.R
\name{pedf}
\alias{pedf}
\title{Adjusted empirical distribution function}
\usage{
pedf(x, data, proper = FALSE)
}
\arguments{
\item{x}{argument of empirical distribution function.}

\item{data}{vector of data for constructing empirical
distribution function.}

\item{proper}{logical variable which when set to TRUE will return
the standard empirical distribution function.}
}
\value{
a vector of same length as x
}
\description{
Adjusted empirical distribution function
}
