% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{pedf}
\alias{pedf}
\title{Empirical Distribution Function}
\usage{
pedf(q, data)
}
\arguments{
\item{q}{vector of values at which to evaluate the empirical distribution
function.}

\item{data}{vector of data values specifying the empirical distribution
function}
}
\value{

}
\description{
Compute a version of the empirical distribution function that is
standardized by (n+1) to lie strictly in (0,1).
}
\examples{
data <- rnorm(100)
pedf(c(-2, -1, 0, 1, 2), data)
}
