% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{plot,marginfit,missing-method}
\alias{plot,marginfit,missing-method}
\title{Plot Method for marginfit Class}
\usage{
\S4method{plot}{marginfit,missing}(x, plotoption = 1L, bw = FALSE)
}
\arguments{
\item{x}{an object of class \linkS4class{marginfit}.}

\item{plotoption}{number of plot required.}

\item{bw}{logical variable specifying whether black-white options should be chosen.}
}
\description{
Plot Method for marginfit Class
}
