% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_models.R
\name{plot,tscmfit,missing-method}
\alias{plot,tscmfit,missing-method}
\title{Plot method for tscmfit class}
\usage{
\S4method{plot}{tscmfit,missing}(x, plottype = "residual", bw = FALSE, lagmax = 30)
}
\arguments{
\item{x}{an object of class \linkS4class{tscmfit}.}

\item{plottype}{type of plot required.}

\item{bw}{logical variable specifying whether black-white options should be chosen.}

\item{lagmax}{maximum lag value for dvinecopula2 plots}
}
\value{
No return value, generates plot.
}
\description{
Plot method for tscmfit class
}
