% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\name{plot,tscopulafit,missing-method}
\alias{plot,tscopulafit,missing-method}
\title{Plot Method for tscopulafit Class}
\usage{
\S4method{plot}{tscopulafit,missing}(x, type = 1L, bw = FALSE)
}
\arguments{
\item{x}{an object of class \linkS4class{tscopulafit}.}

\item{type}{type of plot required.}

\item{bw}{logical variable specifying whether black-white options should be chosen.}

\item{y}{missing.}
}
\value{

}
\description{
Plot Method for tscopulafit Class
}
\examples{
data <- sim(armacopula(list(ar = 0.5, ma = 0.4)), n = 1000)
fit <- fit(armacopula(list(ar = 0.5, ma = 0.4)), data)
plot(fit, type = 1) # try type 1 through 5
}
