% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{pslaplace}
\alias{pslaplace}
\title{Distribution Function of Skew Laplace Distribution}
\usage{
pslaplace(q, mu = 0, scale = 1, gamma = 1)
}
\arguments{
\item{q}{argument of distribution function}

\item{mu}{location parameter}

\item{scale}{scale parameter}

\item{gamma}{skewness parameter}
}
\value{

}
\description{
Distribution Function of Skew Laplace Distribution
}
\keyword{internal}
