% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_vtarma.R
\name{qcondvtarma}
\alias{qcondvtarma}
\title{Conditional Quantiles for vtarma Process}
\usage{
qcondvtarma(p, tscmmod, armamean, armasd)
}
\arguments{
\item{p}{point at which quantile should be calculated.}

\item{tscmmod}{an object of class \linkS4class{tscmfit}.}

\item{armamean}{conditional mean of underlying ARMA process.}

\item{armasd}{conditional standard deviation of underlying ARMA process.}
}
\value{
a scalar value.
}
\description{
Conditional Quantiles for vtarma Process
}
\keyword{internal}
