% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{qmarg}
\alias{qmarg}
\title{Compute quantiles of marginal model}
\usage{
qmarg(x, p)
}
\arguments{
\item{x}{an object of class \linkS4class{margin}.}

\item{p}{vector of probabilities for which quantiles should be computed.}
}
\value{
A vector of values for the quantile function.
}
\description{
Compute the quantile function of the marginal model.
}
\examples{
margmod <- margin("norm", pars = c(mean = 0, sd = 1))
qmarg(margmod, c(0.05, 0.5, 0.95))
}
