% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_vtarma.R
\name{quantile,tscmfit-method}
\alias{quantile,tscmfit-method}
\title{Quantile Calculation Method for tscmfit Class}
\usage{
\S4method{quantile}{tscmfit}(x, alpha, last = FALSE)
}
\arguments{
\item{x}{an object of class \linkS4class{tscmfit}.}

\item{alpha}{a scalar probability value}

\item{last}{logical value asserting that only the last volatility 
prediction should be returned}
}
\value{
a vector of the same length as the data embedded in the tscmfit object.
}
\description{
Quantile Calculation Method for tscmfit Class
}
