% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopulavt.R
\name{RforwardI}
\alias{RforwardI}
\title{Inverse Rosenblatt forward function with v-transforms}
\usage{
RforwardI(x, u, pcs, delta)
}
\arguments{
\item{x}{vector argument of Rosenblatt functiom}

\item{u}{matrix of conditioning values. Number of rows
must be either 1 or same length as x. Number of columns
should not be much more than 15 (due to repeated recursive
calling)}

\item{pcs}{list of pair copulas}

\item{delta}{vector of fulcrum parameters}
}
\value{
vector of same length as x
}
\description{
Inverse Rosenblatt forward function with v-transforms
}
