% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_objects.R
\name{show,swncopula-method}
\alias{show,swncopula-method}
\title{Show Method for Strict White Noise Copula}
\usage{
\S4method{show}{swncopula}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{swncopula}.}
}
\value{
A summary of a \linkS4class{swncopula}.
}
\description{
Show Method for Strict White Noise Copula
}
