% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_models.R
\name{show,tscm-method}
\alias{show,tscm-method}
\title{Show Method for tscm Class}
\usage{
\S4method{show}{tscm}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{tscm}.}
}
\value{

}
\description{
Show Method for tscm Class
}
