% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula.R
\name{sim,dvinecopula-method}
\alias{sim,dvinecopula-method}
\title{Simulation Method for dvinecopula Class}
\usage{
\S4method{sim}{dvinecopula}(x, n = 1000)
}
\arguments{
\item{x}{an object of class \linkS4class{dvinecopula}.}

\item{n}{length of realization.}
}
\value{
A realization of a time series copula process.
}
\description{
Simulation Method for dvinecopula Class
}
\examples{
sim(dvinecopula("gauss", 0.5))
}
