% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_objects.R
\name{sim}
\alias{sim}
\title{New Generic for Simulating Time Series Models}
\usage{
sim(x, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{...}{further arguments to be passed on.}
}
\value{
A realization from the time series model.
}
\description{
Methods are available for objects of class \linkS4class{tscopula},
\linkS4class{margin} and \linkS4class{tsc}.
}
