% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinecopula.R
\name{simdvine}
\alias{simdvine}
\title{D-vine simulation helper function}
\usage{
simdvine(pc_list, n)
}
\arguments{
\item{pc_list}{a list of pair copulas.}

\item{n}{number of data to be simulated.}
}
\value{
a vector of length n.
}
\description{
D-vine simulation helper function
}
