% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacopula.R
\name{starmaStateSpace}
\alias{starmaStateSpace}
\title{State space representation for standardized ARMA model}
\usage{
starmaStateSpace(ar, ma, order)
}
\arguments{
\item{ar}{vector of ar parameters}

\item{ma}{vector of ma parameters}

\item{order}{vector giving order (p,q)}
}
\value{
State space representation of ARMA process.
}
\description{
State space representation for standardized ARMA model
}
\keyword{internal}
