% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_objects.R
\name{swncopula}
\alias{swncopula}
\title{Constructor Function for Strict White Noise Copula}
\usage{
swncopula()
}
\value{
the strict white noise copula process
}
\description{
Constructor Function for Strict White Noise Copula
}
\examples{
swncopula()
}
