% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_vtscopula.R
\name{tsc_objectiveb}
\alias{tsc_objectiveb}
\title{Objective Function for Full Fit With V-Tranform}
\usage{
tsc_objectiveb(
  theta,
  modelspec,
  modeltype,
  dens,
  cdf,
  vt,
  wcopula,
  y,
  fulcrum,
  avoidzero
)
}
\arguments{
\item{theta}{vector of parameters}

\item{modelspec}{list containing model specification}

\item{modeltype}{character vector giving type of model}

\item{dens}{marginal density function}

\item{cdf}{marginal distribution function}

\item{vt}{v-transform}

\item{wcopula}{W copula if used}

\item{y}{vector of data}

\item{fulcrum}{value for fulcrum or NA}

\item{avoidzero}{whether zero should be avoided}
}
\value{

}
\description{
Objective Function for Full Fit With V-Tranform
}
\keyword{internal}
