% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_vtscopula.R
\name{tsc_objectiveb}
\alias{tsc_objectiveb}
\title{Objective function for full fit with v-transform}
\usage{
tsc_objectiveb(theta, fulcrum, modelspec, modeltype, dens, cdf, vt, wcopula, y)
}
\arguments{
\item{theta}{vector of parameters}

\item{fulcrum}{value for fulcrum}

\item{modelspec}{list containing model specification}

\item{modeltype}{character vector giving type of model}

\item{dens}{marginal density function}

\item{cdf}{marginal distribution function}

\item{vt}{v-transform}

\item{wcopula}{W copula if used}

\item{y}{vector of data}
}
\value{
Value of objective function at parameters.
}
\description{
Objective function for full fit with v-transform
}
\keyword{internal}
