% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_models.R
\docType{class}
\name{tscm-class}
\alias{tscm-class}
\title{Full models}
\description{
Class of objects for composite time series models consisting
of stationary copula processes and marginal distributions.
}
\section{Slots}{

\describe{
\item{\code{tscopula}}{an object of class \linkS4class{tscopula}.}

\item{\code{margin}}{an object of class \linkS4class{margin}.}
}}

