% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_models.R
\docType{class}
\name{tscmfit-class}
\alias{tscmfit-class}
\title{Fitted tscm Model}
\value{

}
\description{
Class of objects for fitted \linkS4class{tscm} models.
}
\section{Slots}{

\describe{
\item{\code{tscopula}}{an object of class \linkS4class{tscopula}.}

\item{\code{margin}}{an object of class \linkS4class{margin}.}

\item{\code{data}}{a vector or time series of data to which process has been fitted.}

\item{\code{fit}}{a list containing details of the fit.}
}}

