% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_basic.R
\docType{class}
\name{tscopulafit-class}
\alias{tscopulafit-class}
\title{Fitted Time Series Copula Processes}
\description{
Class of objects for time series copula processes.
}
\section{Slots}{

\describe{
\item{\code{tscopula}}{an object of class \linkS4class{tscopula}.}

\item{\code{data}}{a vector or time series of data.}

\item{\code{fit}}{a list containing details of the fit.}
}}

