% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtransforms.R
\name{Vdegenerate}
\alias{Vdegenerate}
\title{Constructor function for degenerate v-transform}
\usage{
Vdegenerate()
}
\value{
An object of class \linkS4class{VtransformI}.
}
\description{
Constructor function for degenerate v-transform
}
\examples{
Vdegenerate()
}
