% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtransforms.R
\name{vdownprob}
\alias{vdownprob}
\title{Calculate conditional down probability of v-transform}
\usage{
vdownprob(x, v)
}
\arguments{
\item{x}{an object of class \linkS4class{Vtransform}.}

\item{v}{a vector or time series with values in [0, 1].}
}
\value{
A vector or time series of values of gradient.
}
\description{
Calculate conditional down probability of v-transform
}
\examples{
vdownprob(V2p(delta = 0.55, kapp = 1.2), c(0, 0.25, 0.5, 0.75, 1))
}
