% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtransforms.R
\name{vgradient}
\alias{vgradient}
\title{Calculate Gradient of V-Transform}
\usage{
vgradient(x, u)
}
\arguments{
\item{x}{a \code{Vtransform} object.}

\item{u}{a vector or time series with values in [0, 1].}
}
\value{
A vector or time series of values of gradient.
}
\description{
Calculate Gradient of V-Transform
}
\examples{
vgradient(Vsymmetric(), c(0, 0.25, 0.5, 0.75, 1))
}
