% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtransforms.R
\name{vtrans}
\alias{vtrans}
\title{Evaluate a V-Transform}
\usage{
vtrans(x, u, correction = FALSE)
}
\arguments{
\item{x}{a \code{Vtransform} object.}

\item{u}{a vector or time series with values in [0, 1].}

\item{correction}{logical variable specifying whether v-transform
should be bounded away from zero.}
}
\value{
A vector or time series with values in [0, 1].
}
\description{
Evaluate a V-Transform
}
\examples{
vtrans(Vsymmetric(), c(0, 0.25, 0.5, 0.75, 1))
}
