% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtransforms.R
\docType{class}
\name{Vtransform-class}
\alias{Vtransform-class}
\title{V-Transform Class}
\value{

}
\description{
V-Transform Class
}
\section{Slots}{

\describe{
\item{\code{name}}{a name for the v-transform of class character.}

\item{\code{Vtrans}}{function to evaluate the v-transform.}

\item{\code{pars}}{vector containing the named parameters of the v-transform.}

\item{\code{gradient}}{function to evaluate the gradient of the v-transform.}
}}

\examples{
V2p(delta = 0.5, kappa = 1.2)
}
