% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_vtscopula.R
\name{vtscopula}
\alias{vtscopula}
\title{Constructor function for vtscopula object}
\usage{
vtscopula(tscopulaU, Vtransform = Vlinear(), Wcopula = swncopula())
}
\arguments{
\item{tscopulaU}{an object of class
\linkS4class{armacopula}, \linkS4class{dvinecopula} or \linkS4class{dvinecopula2}.}

\item{Vtransform}{an object of class \linkS4class{Vtransform}.}

\item{Wcopula}{an object of class \linkS4class{tscopula}.}
}
\value{
An object of class \linkS4class{vtscopula}.
}
\description{
Constructor function for vtscopula object
}
\examples{
copobject <- armacopula(pars = list(ar = 0.6, ma = 0.2))
vtscopula(copobject, Vtransform = V2p())
}
