% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_vtscopula.R
\name{vtscopula_objective}
\alias{vtscopula_objective}
\title{Objective function for gvtscopula fitting}
\usage{
vtscopula_objective(
  theta,
  modelspec,
  modeltype,
  vt,
  wcopula,
  y,
  fulcrum,
  avoidzero
)
}
\arguments{
\item{theta}{vector of parameters}

\item{modelspec}{list containing tscopula specification}

\item{modeltype}{character vector specifying tscopula type}

\item{vt}{a v transform}

\item{wcopula}{W copula for serial dependence}

\item{y}{vector of data}

\item{fulcrum}{fixed value for fulcrum or NA}

\item{avoidzero}{logical for whether zero should be avoided}
}
\value{

}
\description{
Objective function for gvtscopula fitting
}
\keyword{internal}
